Move MainWindow member functions to external source files in JS8_Mainwindow: - buildQueryMap.cpp - checkVersion.cpp - dataSink.cpp - displayBandActivity.cpp - displayCallActivity.cpp - initializeDummyDate.cpp - initializeGroupMessage.cpp - networkMessage.cpp - processCommandActivity.cpp - processDecodeEvent.cpp - processRxActivity.cpp
reduces mainwindow.cpp from 12,482 to 8,284 lines of code
