+++ JS8_UI/mainwindow.cpp
@@ -1,90 +1,7 @@
 //---------------------------------------------------------- MainWindow
 #include "mainwindow.h"
-#include <QAction>
-#include <QActionGroup>
-#include <QApplication>
-#include <QByteArrayView>
-#include <QCursor>
-#include <QDesktopServices>
-#include <QDir>
-#include <QElapsedTimer>
-#include <QFileDialog>
-#include <QHostInfo>
-#include <QInputDialog>
-#include <QLineEdit>
-#include <QLoggingCategory>
-#include <QMdiSubWindow>
-#include <QNetworkAccessManager>
-#include <QNetworkReply>
-#include <QNetworkRequest>
-#include <QPixmap>
-#include <QProgressDialog>
-#include <QRegularExpression>
-#include <QRegularExpressionValidator>
-#include <QScrollBar>
-#include <QSoundEffect>
-#include <QStandardPaths>
-#include <QStringBuilder>
-#include <QTime>
-#include <QTimeZone>
-#include <QToolTip>
-#include <QUdpSocket>
-#include <QUrl>
-#include <QVariant>
-#include <QVector>
-#include <QVersionNumber>
-#include <QtConcurrent/QtConcurrentRun>
-#include <algorithm>
-#include <array>
-#include <boost/crc.hpp>
-#include <cmath>
-#include <complex>
-#include <cstring>
-#include <fftw3.h>
-#include <functional>
-#include <iterator>
-#include <mutex>
-#include <stdexcept>
-#include <string_view>
-#include <vector>
-
-#include "JS8_Audio/NotificationAudio.h"
-#include "JS8_Audio/soundin.h"
-#include "JS8_Audio/soundout.h"
-#include "JS8_Include/EventFilter.h"
-#include "JS8_Main/Bands.h"
-#include "JS8_Main/DriftingDateTime.h"
-#include "JS8_Main/Geodesic.h"
-#include "JS8_Main/HelpTextWindow.h"
-#include "JS8_Main/Inbox.h"
-#include "JS8_Main/MessageClient.h"
-#include "JS8_Main/MultiSettings.h"
-#include "JS8_Main/Radio.h"
-#include "JS8_Main/SelfDestructMessageBox.h"
-#include "JS8_Main/SignalMeter.h"
-#include "JS8_Main/StationList.h"
-#include "JS8_Main/qt_helpers.h"
-#include "JS8_Main/revision_utils.h"
-#include "JS8_Mode/Decoder.h"
-#include "JS8_Mode/Detector.h"
-#include "JS8_Mode/JS8Submode.h"
-#include "JS8_Mode/Modulator.h"
-#include "JS8_Mode/decodedtext.h"
-#include "JS8_Network/NetworkAccessManager.h"
-#include "JS8_Transceiver/TransceiverFactory.h"
-#include "JS8_UDP/WSJTXMessageClient.h"
-#include "JS8_UDP/WSJTXMessageMapper.h"
-#include "JS8_UI/about.h"
-#include "JS8_UI/widegraph.h"
-#include "JS8_jsc/jsc_checker.h"
-#include "logqso.h"
-#include "messagereplydialog.h"
-#include "messagewindow.h"
 
 #include "moc_mainwindow.cpp"
-#include "ui_mainwindow.h"
-
-Q_DECLARE_LOGGING_CATEGORY(mainwindow_js8)
 
 // TODO: Move to member:
 static char message[29];
@@ -122,21 +39,6 @@ int ms_minute_error() {
 
     return now.msecsTo(now.addSecs(second > 30 ? 60 - second : -second)) -
            time.msec();
-}
-
-QString since(QDateTime const &time) {
-    auto const delta = time.secsTo(DriftingDateTime::currentDateTimeUtc());
-
-    if (delta >= 60 * 60 * 24)
-        return QString("%1d").arg(delta / (60 * 60 * 24));
-    else if (delta >= 60 * 60)
-        return QString("%1h").arg(delta / (60 * 60));
-    else if (delta >= 60)
-        return QString("%1m").arg(delta / (60));
-    else if (delta >= 15)
-        return QString("%1s").arg(delta - (delta % 15));
-    else
-        return QString("now");
 }
 
 namespace State {
@@ -184,18 +86,6 @@ int roundUp(int numToRound, int multiple) {
     return roundDown + multiple;
 }
 
-auto replaceMacros(QString const &text, QMap<QString, QString> const &values,
-                   bool const prune) {
-    QString output = text;
-
-    for (auto const [key, value] : values.asKeyValueRange()) {
-        output = output.replace(key, value.toUpper());
-    }
-
-    return prune ? output.replace(QRegularExpression("[<](?:[^>]+)[>]"), "")
-                 : output;
-}
-
 // Copy at most size bytes into the array, filling any unused size
 // with spaces if less than size bytes were available to copy. For
 // convenience, return a one past the end iterator, i.e., equal to
@@ -218,72 +108,6 @@ void copyMessage(QStringView const string, char *const array,
     *copyByteData(string.toLocal8Bit(), array, size) = '\0';
 }
 
-// Emulation of the Fortran 'flat1' subroutine.
-
-void flat1(float const *const savg, int const iz, int const nsmo,
-           float *const slin) {
-    constexpr int x_size = 8192;
-    constexpr int nstep = 20;
-    constexpr int nh = nstep / 2;
-
-    // Define bounds for smoothing
-    int const ia = nsmo / 2 + 1;
-    int const ib = iz - nsmo / 2 - 1;
-
-    std::vector<float> x(x_size, 0.0f);
-
-    // Smooth savg using median percentiles
-
-    auto const rank =
-        std::clamp(static_cast<int>(std::round(0.5f * nsmo)), 0, nsmo - 1);
-
-    for (int i = ia; i <= ib; i += nstep) {
-        auto const data = &savg[i - nsmo / 2];
-        auto temp = std::vector<float>(data, data + nsmo);
-
-        std::nth_element(temp.begin(), temp.begin() + rank, temp.end());
-
-        x[i] = temp[rank];
-
-        std::fill(x.begin() + (i - nh), x.begin() + (i + nh), x[i]);
-    }
-
-    // Extend smoothed values to boundaries
-    std::fill(x.begin(), x.begin() + ia, x[ia]);
-    std::fill(x.begin() + ib + 1, x.begin() + iz, x[ib]);
-
-    // Compute scaling factor
-    float x0 = 0.001f * *std::max_element(x.begin() + iz / 10,
-                                          x.begin() + (9 * iz) / 10);
-
-    // Normalize savg to compute slin
-    for (int i = 0; i < iz; ++i)
-        slin[i] = savg[i] / (x[i] + x0);
-}
-
-// Emulation of the Fortran 'smo' subroutine. However, doesn't copy the data
-// back from b to a; rather, a is input and, b is output. Since we invariably
-// call this twice, we can just swap the order of the arrays to achieve the
-// same result without the extra two copy operations.
-
-void smo(float const *const a, float *const b, int const npts, int const nadd) {
-    auto const nh = nadd / 2;
-
-    // Smooth the array
-    for (int i = nh; i < npts - nh; ++i) {
-        float sum = 0.0f;
-        for (int j = -nh; j <= nh; ++j) {
-            sum += a[i + j];
-        }
-        b[i] = sum;
-    }
-
-    // Set edges to zero
-    for (int i = 0; i < nh; ++i)
-        b[i] = 0.0f; // Zero out leading edge
-    for (int i = npts - nh; i < npts; ++i)
-        b[i] = 0.0f; // Zero out trailing edge
-}
 } // namespace
 
 //--------------------------------------------------- MainWindow constructor
@@ -1659,66 +1483,14 @@ MainWindow::MainWindow(QString const &program_info, QDir const &temp_directory,
         << "Initialization with onPlumbingCompleted() has completed.";
 
     QTimer::singleShot(500, this, &MainWindow::initializeDummyData);
-    QTimer::singleShot(500, this, &MainWindow::initializeGroupMessageDummyData);
+    QTimer::singleShot(500, this, &MainWindow::initializeGroupMessage);
 
     // this must be the last statement of constructor
     if (!m_valid)
         throw std::runtime_error{"Fatal initialization exception"};
 }
 
-void MainWindow::checkVersion(bool const alertOnUpToDate) {
-    auto m = new QNetworkAccessManager(this);
-    connect(m, &QNetworkAccessManager::finished, this,
-            [this, alertOnUpToDate](QNetworkReply *reply) {
-                if (reply->error()) {
-                    qCDebug(mainwindow_js8) << "Checking for Updates Error:"
-                                            << reply->errorString();
-                    return;
-                }
-
-                QString content = reply->readAll().trimmed();
-
-                auto const currentVersion =
-                    QVersionNumber::fromString(version());
-                auto const networkVersion = QVersionNumber::fromString(content);
-
-                qCDebug(mainwindow_js8) << "Checking Version" << currentVersion
-                                        << "with" << networkVersion;
-
-                if (currentVersion < networkVersion) {
-
-                    SelfDestructMessageBox *m = new SelfDestructMessageBox(
-                        60, "New Updates Available",
-                        QString("A new version (%1) of JS8Call is now "
-                                "available. Please see the <a "
-                                "href='https://github.com/JS8Call-improved/"
-                                "JS8Call-improved/releases'>GitHub "
-                                "Releases</a> for more details.")
-                            .arg(content),
-                        QMessageBox::Information, QMessageBox::Ok,
-                        QMessageBox::Ok, false, this);
-
-                    m->show();
-
-                } else if (alertOnUpToDate) {
-
-                    SelfDestructMessageBox *m = new SelfDestructMessageBox(
-                        60, "No Updates Available",
-                        QString("Your version (%1) of JS8Call is up-to-date.")
-                            .arg(version()),
-                        QMessageBox::Information, QMessageBox::Ok,
-                        QMessageBox::Ok, false, this);
-
-                    m->show();
-                }
-            });
-
-    qCDebug(mainwindow_js8) << "Checking for Updates...";
-    QUrl url("https://github.com/JS8Call-improved/JS8Call-improved/releases/"
-             "latest/download/version.txt");
-    QNetworkRequest r(url);
-    m->get(r);
-}
+bool checkVersion(); // JS8_Mainwindow/checkVersion.cpp
 
 void MainWindow::checkStartupWarnings() {
     if (m_config.check_for_updates()) {
@@ -1727,424 +1499,9 @@ void MainWindow::checkStartupWarnings() {
     ensureCallsignSet(false);
 }
 
-void MainWindow::initializeDummyData() {
-    if (!QApplication::applicationName().contains("dummy")) {
-        return;
-    }
+void initializeDummyData(); // JS8_Mainwindow/initializeDummyData.cpp
 
-#if 0
-    Codeword all;
-    foreach(CodewordPair p, JSC::compress("")){
-        all.append(p.first);
-    }
-    qCDebug(mainwindow_js8) << all;
-    qCDebug(mainwindow_js8) << JSC::decompress(all) << (JSC::decompress(all) == "HELLO WORLD ");
-    exit(-1);
-#endif
-
-#if 0
-    auto t = new QTimer(this);
-    t->setInterval(150);
-    connect(t, &QTimer::timeout, this, [this](){
-        if(!ui->extFreeTextMsgEdit->hasFocus()){
-            return;
-        }
-
-        auto c = ui->extFreeTextMsgEdit->textCursor();
-        int pos = qMin(c.selectionStart(), c.selectionEnd());
-
-        if(pos <= 5 && c.selectionStart() != c.selectionEnd()){
-            c.clearSelection();
-            ui->extFreeTextMsgEdit->setTextCursor(c);
-        }
-    });
-    t->start();
-
-    auto kpe = new KeyPressEater();
-    connect(kpe, &KeyPressEater::keyPressed, this, [this](QObject *obj, QKeyEvent * e, bool *pProcessed){
-        auto t = e->text();
-        auto c = ui->extFreeTextMsgEdit->textCursor();
-        int pos = qMin(c.selectionStart(), c.selectionEnd());
-
-        if(e->key() == Qt::Key_Escape){
-            *pProcessed = false;
-            return;
-        }
-
-        QTextCursor cc(c);
-        cc.setPosition(5);
-        cc.movePosition(QTextCursor::NextWord);
-        int cpos = qMax(cc.selectionStart(), cc.selectionEnd());
-        if(e->key() == Qt::Key_Backspace && e->modifiers().testFlag(Qt::ControlModifier) && pos < cpos){
-            *pProcessed = true;
-            return;
-        }
-
-        if((e->key() == Qt::Key_Backspace && pos <= 5) ||
-           (e->key() == Qt::Key_Delete && pos < 5)){
-            *pProcessed = true;
-            return;
-        }
-
-        if(e->key() == Qt::Key_V && e->modifiers().testFlag(Qt::ControlModifier) && pos <= 5){
-            *pProcessed = true;
-            return;
-        }
-
-        if(!t.isEmpty() && pos < 5){
-            *pProcessed = true;
-            return;
-        }
-    });
-    ui->extFreeTextMsgEdit->installEventFilter(kpe);
-
-    connect(ui->extFreeTextMsgEdit, &QTextEdit::copyAvailable, this, [this](bool available){
-        if(!available){
-            return;
-        }
-        auto c = ui->extFreeTextMsgEdit->textCursor();
-
-        qCDebug(mainwindow_js8) << "select" << c.selectionStart() << c.selectionEnd();
-
-        int pos = qMin(c.selectionStart(), c.selectionEnd());
-        if(pos <= 5){
-            auto text = c.selectedText();
-            if(!text.isEmpty()){
-                c.clearSelection();
-                ui->extFreeTextMsgEdit->setTextCursor(c);
-            }
-        }
-    });
-
-    /*
-    connect(ui->extFreeTextMsgEdit->document(), &QTextDocument::cursorPositionChanged, this, [this](const QTextCursor &){
-        auto c = ui->extFreeTextMsgEdit->textCursor();
-        int pos = qMin(c.selectionStart(), c.selectionEnd());
-        if(pos <= 5){
-            auto text = c.selectedText();
-            if(!text.isEmpty()){
-                c.clearSelection();
-                ui->extFreeTextMsgEdit->setTextCursor(c);
-            }
-        }
-    });
-
-    connect(ui->extFreeTextMsgEdit, &QTextEdit::cursorPositionChanged, this, [this](){
-        auto c = ui->extFreeTextMsgEdit->textCursor();
-        int pos = qMin(c.selectionStart(), c.selectionEnd());
-
-        if(pos <= 5){
-            auto text = c.selectedText();
-            if(!text.isEmpty()){
-                c.clearSelection();
-                ui->extFreeTextMsgEdit->setTextCursor(c);
-            }
-        }
-    });
-    */
-
-    /*connect(ui->extFreeTextMsgEdit->document(), &QTextDocument::contentsChange, this, [this](int from, int removed, int added){
-        if(from < 5 && removed == 1){
-            ui->extFreeTextMsgEdit->document()->blockSignals(true);
-            ui->extFreeTextMsgEdit->document()->undo();
-            ui->extFreeTextMsgEdit->document()->clearUndoRedoStacks(QTextDocument::RedoStack);
-            ui->extFreeTextMsgEdit->document()->blockSignals(false);
-        }
-    });*/
-
-    ui->extFreeTextMsgEdit->setText("HELLO BRAVE NEW WORLD");
-    auto c = ui->extFreeTextMsgEdit->textCursor();
-    c.setPosition(0);
-    c.setPosition(5, QTextCursor::KeepAnchor);
-    auto f = c.charFormat();
-    f.setFontStrikeOut(true);
-    c.setCharFormat(f);
-#endif
-
-    // this causes a segfault!
-    // processDecodedLine("223000 -15 -0.3 1681 B  6t++yk+aJbaE         6 \n");
-
-    ui->extFreeTextMsgEdit->setPlainText("HELLOBRAVE NEW WORLD");
-    ui->extFreeTextMsgEdit->setCharsSent(6);
-
-    logHeardGraph("KN4CRD", "OH8STN");
-    logHeardGraph("KN4CRD", "K0OG");
-    logHeardGraph("K0OG", "KN4CRD");
-
-    auto path = QDir::toNativeSeparators(
-        m_config.writeable_data_dir().absoluteFilePath(QString("test.db3")));
-    auto inbox = Inbox(path);
-    if (inbox.open()) {
-        qCDebug(mainwindow_js8) << "test inbox opened"
-                                << inbox.count("test", "$", "%") << "messages";
-
-        // int i = inbox.append(Message("test", "booya1"));
-
-        int i = inbox.append(Message("test", "booya2"));
-        qCDebug(mainwindow_js8) << "i" << i;
-
-        qCDebug(mainwindow_js8) << inbox.set(i, Message("test", "booya3"));
-
-        auto m = inbox.value(i);
-        qCDebug(mainwindow_js8) << QString(m.toJson());
-
-        qCDebug(mainwindow_js8) << inbox.del(i);
-
-        foreach (auto pair, inbox.values("test", "$", "%", 0, 5)) {
-            qCDebug(mainwindow_js8)
-                << pair.first << QString(pair.second.toJson());
-        }
-    }
-
-    auto d = DecodedText("SN5-lUuJkby0", Varicode::JS8CallFirst, 1);
-    qCDebug(mainwindow_js8) << "KN4CRD: K0OG ===>" << d.message();
-
-    // qCDebug(mainwindow_js8) << Varicode::isValidCallsign("@GROUP1", nullptr);
-    // qCDebug(mainwindow_js8) << Varicode::packAlphaNumeric50("VE7/KN4CRD");
-    // qCDebug(mainwindow_js8) <<
-    // Varicode::unpackAlphaNumeric50(Varicode::packAlphaNumeric50("VE7/KN4CRD"));
-    // qCDebug(mainwindow_js8) <<
-    // Varicode::unpackAlphaNumeric50(Varicode::packAlphaNumeric50("@GROUP/42"));
-    // qCDebug(mainwindow_js8) <<
-    // Varicode::unpackAlphaNumeric50(Varicode::packAlphaNumeric50("SP1ATOM"));
-
-    if (!m_config.my_groups().contains("@GROUP42")) {
-        m_config.addGroup("@GROUP42");
-    }
-
-    QList<QString> calls = {"KN4CRD", "VE7/KN4CRD", "KN4CRD/P", "KC9QNE",
-                            "KI6SSI", "K0OG",       "LB9YH",    "VE7/LB9YH",
-                            "M0IAX",  "N0JDS",      "OH8STN",   "VA3OSO",
-                            "VK1MIC", "W0FW"};
-
-    auto dt = DriftingDateTime::currentDateTimeUtc().addSecs(-300);
-
-    int i = 0;
-    foreach (auto call, calls) {
-        CallDetail cd = {};
-        cd.call = call;
-        cd.through = i == 2 ? "KN4CRD" : "";
-        cd.dial = 7078000;
-        cd.offset = 500 + 100 * i;
-        cd.snr = i == 3 ? -100 : i;
-        cd.ackTimestamp = i == 1 ? dt.addSecs(-900) : QDateTime{};
-        cd.utcTimestamp = dt;
-        cd.grid = i == 5 ? "J042" : i == 6 ? " FN42FN42FN" : "";
-        cd.tdrift = 0.1 * i;
-        cd.submode = i % 3;
-        logCallActivity(cd, false);
-
-        ActivityDetail ad = {};
-        ad.bits = Varicode::JS8CallFirst | Varicode::JS8CallLast;
-        ad.snr = i == 3 ? -100 : i;
-        ad.dial = 7078000;
-        ad.offset = 500 + 100 * i;
-        ad.text = QString("%1: %2 TEST MESSAGE")
-                      .arg(call)
-                      .arg(m_config.my_callsign());
-        ad.utcTimestamp = dt;
-        ad.submode = cd.submode;
-        m_bandActivity[500 + 100 * i] = {ad};
-
-        markOffsetDirected(500 + 100 * i, false);
-
-        i++;
-    }
-
-    ActivityDetail adHB1 = {};
-    adHB1.bits = Varicode::JS8CallFirst;
-    adHB1.snr = 0;
-    adHB1.dial = 7078000;
-    adHB1.offset = 750;
-    adHB1.text = QString("KN4CRD: HB AUTO EM73");
-    adHB1.utcTimestamp = DriftingDateTime::currentDateTimeUtc();
-    adHB1.submode = Varicode::JS8CallNormal;
-    m_bandActivity[750].append(adHB1);
-
-    ActivityDetail adHB2 = {};
-    adHB2.bits = Varicode::JS8CallLast;
-    adHB2.snr = 0;
-    adHB2.dial = 7078000;
-    adHB2.offset = 750;
-    adHB2.text = QString(" MSG ID 1");
-    adHB2.utcTimestamp = DriftingDateTime::currentDateTimeUtc();
-    adHB2.submode = Varicode::JS8CallNormal;
-    m_bandActivity[750].append(adHB2);
-
-    CommandDetail cmd = {};
-    cmd.cmd = ">";
-    cmd.to = m_config.my_callsign();
-    cmd.from = "N0JDS";
-    cmd.relayPath = "N0JDS>OH8STN";
-    cmd.text = "HELLO BRAVE SOUL";
-    cmd.utcTimestamp = dt;
-    cmd.submode = Varicode::JS8CallNormal;
-    addCommandToMyInbox(cmd);
-
-    QString eot = m_config.eot();
-
-    displayTextForFreq(QString("KN4CRD: @ALLCALL? %1 ").arg(eot), 42,
-                       DriftingDateTime::currentDateTimeUtc().addSecs(-315),
-                       true, true, true);
-    displayTextForFreq(QString("J1Y: KN4CRD SNR -05 %1 ").arg(eot), 42,
-                       DriftingDateTime::currentDateTimeUtc().addSecs(-300),
-                       false, true, true);
-    displayTextForFreq(QString("HELLO BRAVE  NEW   WORLD    %1 ").arg(eot), 42,
-                       DriftingDateTime::currentDateTimeUtc().addSecs(-300),
-                       false, true, true);
-
-    auto now = DriftingDateTime::currentDateTimeUtc();
-    displayTextForFreq(QString("KN4CRD: JY1 ACK -12 %1 ").arg(eot), 780, now,
-                       false, true, true);
-    displayTextForFreq(QString("KN4CRD: JY1 ACK -12 %1 ").arg(eot), 780, now,
-                       false, true, true); // should be hidden (duplicate)
-    displayTextForFreq(QString("OH8STN: JY1 ACK -12 %1 ").arg(eot), 780, now,
-                       false, true, true);
-
-    displayTextForFreq(QString("KN4CRD: JY1 ACK -10 %1 ").arg(eot), 800, now,
-                       false, true, true);
-    displayTextForFreq(QString("KN4CRD: JY1 ACK -12 %1 ").arg(eot), 780,
-                       now.addSecs(120), false, true, true);
-
-    displayTextForFreq(QString("HELLO\\nBRAVE\\nNEW\\nWORLD %1 ").arg(eot),
-                       1500, now, false, true, true);
-
-    displayActivity(true);
-}
-
-void MainWindow::initializeGroupMessageDummyData() {
-    if (!QApplication::applicationName().contains("dummy")) {
-        return;
-    }
-
-    if (!m_config.my_groups().contains("@GROUP42")) {
-        m_config.addGroup("@GROUP42");
-    }
-
-    auto dt = DriftingDateTime::currentDateTimeUtc().addSecs(-300);
-
-    CommandDetail cmd = {};
-    cmd.cmd = " MSG TO:";
-    cmd.from = "KN4CRD";
-    cmd.to = "@GROUP42";
-    cmd.utcTimestamp = dt;
-    cmd.submode = Varicode::JS8CallNormal;
-    cmd.text = "@GROUP42 TEST MESSAGE TO GROUP";
-
-    m_rxCommandQueue.append(cmd);
-
-    processCommandActivity();
-
-    dt = DriftingDateTime::currentDateTimeUtc().addSecs(-150);
-
-    CommandDetail cmd1 = {};
-    cmd1.cmd = " MSG TO:";
-    cmd1.from = "KN4CRD";
-    cmd1.to = "@GROUP42";
-    cmd1.utcTimestamp = dt;
-    cmd1.submode = Varicode::JS8CallNormal;
-    cmd1.text = "@GROUP42 ANOTHER TEST MESSAGE TO GROUP";
-
-    m_rxCommandQueue.append(cmd1);
-
-    processCommandActivity();
-
-    CommandDetail cmd2 = {};
-    cmd2.cmd = " QUERY MSGS";
-    cmd2.from = "W1AW";
-    cmd2.to = "@GROUP42";
-    cmd2.utcTimestamp = DriftingDateTime::currentDateTimeUtc();
-    cmd2.submode = Varicode::JS8CallNormal;
-
-    m_rxCommandQueue.append(cmd2);
-
-    processCommandActivity();
-
-    int mid = getNextGroupMessageIdForCallsign("@GROUP42", "W1AW");
-
-    qCDebug(mainwindow_js8) << "Testing group messaging";
-    qCDebug(mainwindow_js8) << "Test message ID: " << mid;
-
-    std::string textString = "MSG ";
-    textString += std::to_string(mid);
-
-    qCDebug(mainwindow_js8) << "Text string: " << textString.c_str();
-
-    CommandDetail cmd3 = {};
-    cmd3.cmd = " QUERY";
-    cmd3.from = "W1AW";
-    cmd3.to = "@GROUP42";
-    cmd3.utcTimestamp = DriftingDateTime::currentDateTimeUtc();
-    cmd3.submode = Varicode::JS8CallNormal;
-    cmd3.text = textString.c_str();
-
-    m_rxCommandQueue.append(cmd3);
-
-    processCommandActivity();
-
-    dt = DriftingDateTime::currentDateTimeUtc().addSecs(-300);
-
-    CommandDetail cmd4 = {};
-    cmd4.cmd = " MSG TO:";
-    cmd4.from = "KN4CRD";
-    cmd4.to = "K4RWR";
-    cmd4.utcTimestamp = dt;
-    cmd4.submode = Varicode::JS8CallNormal;
-    cmd4.text = "W1AW TEST MESSAGE TO STATION";
-
-    m_rxCommandQueue.append(cmd4);
-
-    processCommandActivity();
-
-    dt = DriftingDateTime::currentDateTimeUtc().addSecs(-150);
-
-    CommandDetail cmd5 = {};
-    cmd5.cmd = " MSG TO:";
-    cmd5.from = "KN4CRD";
-    cmd5.to = "K4RWR";
-    cmd5.utcTimestamp = dt;
-    cmd5.submode = Varicode::JS8CallNormal;
-    cmd5.text = "W1AW ANOTHER TEST MESSAGE TO STATION";
-
-    m_rxCommandQueue.append(cmd5);
-
-    processCommandActivity();
-
-    CommandDetail cmd6 = {};
-    cmd6.cmd = " QUERY MSGS";
-    cmd6.from = "W1AW";
-    cmd6.to = "K4RWR";
-    cmd6.utcTimestamp = DriftingDateTime::currentDateTimeUtc();
-    cmd6.submode = Varicode::JS8CallNormal;
-
-    m_rxCommandQueue.append(cmd6);
-
-    processCommandActivity();
-
-    mid = getNextMessageIdForCallsign("W1AW");
-
-    qCDebug(mainwindow_js8) << "Testing group messaging";
-    qCDebug(mainwindow_js8) << "Test message ID: " << mid;
-
-    textString = "MSG ";
-    textString += std::to_string(mid);
-
-    qCDebug(mainwindow_js8) << "Text string: " << textString.c_str();
-
-    CommandDetail cmd7 = {};
-    cmd7.cmd = " QUERY";
-    cmd7.from = "W1AW";
-    cmd7.to = "K4RWR";
-    cmd7.utcTimestamp = DriftingDateTime::currentDateTimeUtc();
-    cmd7.submode = Varicode::JS8CallNormal;
-    cmd7.text = textString.c_str();
-
-    m_rxCommandQueue.append(cmd7);
-
-    processCommandActivity();
-
-    displayActivity(true);
-}
+void initializeGroupMessage(); // JS8_Mainwindow/initializeGroupMessage.cpp
 
 void MainWindow::initialize_fonts() {
     set_application_font(m_config.text_font());
@@ -2625,197 +1982,7 @@ void MainWindow::set_application_font(QFont const &font) {
     }
 }
 
-//-------------------------------------------------------------- dataSink()
-void MainWindow::dataSink(qint64 frames) {
-    constexpr int NMAX = JS8_NTMAX * 12000;
-    constexpr int nfft3 = 16384;
-    constexpr std::array nch = {1, 2, 4, 9, 18, 36, 72};
-
-    // symspec global vars
-    static int ja = 0;
-    static int k0 = 999999999;
-    static WF::SPlot ssum = {};
-    static WF::SPlot s = {};
-
-    int k(frames);
-    if (k0 == 999999999) {
-        m_ihsym = int((float)frames / (float)JS8_NSPS) * 2;
-        ja = k;
-        k0 = k;
-    }
-
-    // qCDebug(mainwindow_js8) << "k" << k << "k0" << k0 << "delta" << k-k0;
-
-    // Get power, spectrum, and ihsym
-    int const nsmo = m_wideGraph->smoothYellow() - 1;
-
-    // make sure the ssum global is reset every period cycle
-    static int lastCycle = -1;
-    int const cycle = JS8::Submode::computeCycleForDecode(m_nSubMode, k);
-
-    if (cycle != lastCycle) {
-        qCDebug(decoder_js8) << "period loop, resetting ssum";
-        ssum.fill(0.0f);
-    }
-
-    lastCycle = cycle;
-
-    // cap ihsym based on the period max
-    m_ihsym = m_ihsym % (m_TRperiod * JS8_RX_SAMPLE_RATE / JS8_NSPS * 2);
-
-    int const jstep = JS8_NSPS / 2;
-
-    if (k >= 2048 && k <= NMAX) {
-        if (k < k0) {
-            // Start a new data block
-            ja = 0;
-            ssum.fill(0.0f);
-            m_ihsym = 0;
-            std::fill(std::begin(dec_data.d2) + k, std::end(dec_data.d2), 0);
-        }
-
-        float gain = pow(10.0f, 0.1f * m_inGain);
-        float sq = 0.0f;
-        float pxmax = 0.0f;
-
-        for (int i = k0; i < k; ++i) {
-            float x1 = dec_data.d2[i];
-            pxmax = std::max(pxmax, fabs(x1));
-            sq += x1 * x1;
-        }
-
-        m_px = sq > 0.0f ? 10.0f * log10(sq / (k - k0)) : 0.0f;
-        m_pxmax = pxmax > 0.0f ? 20.0f * log10(pxmax) : 0.0f;
-
-        k0 = k;
-        ja += jstep;
-
-        // Perform real to complex FFT; note that the Fortran code performed
-        // this operation in the four2a subroutine, which contains a cache
-        // of plans. However, since the FFTW3 library itself caches, this
-        // resulted in double caching, so as an experiment we're allowing
-        // the library to handle caching.
-        //
-        // Note that the FFTW library requires all calls but for the plan
-        // execution, i.e., fftwf_execute(), to be serialized. The library
-        // doesn't require prescriptive serialization, anything will do so
-        // long as the result is one thread at a time.
-        //
-        // Providing room for an extra complex value, i.e., a pair of floats,
-        // real and imaginary parts, allows us to use the same buffer for the
-        // FFT input and output.
-        //
-        // While the memory for the FFT can come from anywhere, if we ask the
-        // library for it, it'll guarantee that it's aligned for use of SIMD
-        // instructions, which will in turn allow it to use them.
-
-        fftwf_complex *fftw_complex;
-        float *fftw_real;
-        fftwf_plan fftw_plan;
-        {
-            std::lock_guard<std::mutex> lock(fftw_mutex);
-
-            fftw_complex = fftwf_alloc_complex(nfft3 / 2 + 1);
-
-            if (!fftw_complex) {
-                throw std::runtime_error("Failed to allocate FFT data");
-            }
-
-            fftw_real = reinterpret_cast<float *>(fftw_complex);
-            fftw_plan = fftwf_plan_dft_r2c_1d(nfft3, fftw_real, fftw_complex,
-                                              FFTW_ESTIMATE_PATIENT);
-
-            if (!fftw_plan) {
-                fftwf_free(fftw_complex);
-                throw std::runtime_error("Failed to create FFT plan");
-            }
-        }
-
-        // Copy data and apply the window, then execute the FFT.
-
-        for (int i = 0; i < nfft3; ++i) {
-            if (int j = ja + i - nfft3; j >= 0 && j < NMAX)
-                fftw_real[i] = 0.1f * dec_data.d2[j];
-        }
-
-        ++m_ihsym;
-
-        fftwf_execute(fftw_plan);
-
-        // Process the resulting spectrum.
-
-        m_df3 = 12000.0f / nfft3;
-
-        auto const iz = std::min(JS8_NSMAX, static_cast<int>(5000.0f / m_df3));
-        auto const cx = reinterpret_cast<std::complex<float> *>(fftw_complex);
-        auto const fac = std::pow(1.0f / nfft3, 2.0f);
-
-        for (int i = 0; i < iz; ++i) {
-            auto const sx = fac * std::norm(cx[i]);
-            ssum[i] += sx;
-            s[i] = 1000.0f * gain * sx;
-        }
-
-        // We're now done with the plan and the FFT storage. Even though
-        // the plan is being 'destroyed' here, the library will cache it
-        // at its discretion. As above, the calls must be serialized.
-
-        {
-            std::lock_guard<std::mutex> lock(fftw_mutex);
-            fftwf_destroy_plan(fftw_plan);
-            fftwf_free(fftw_complex);
-        }
-
-        // Update average spectra.
-
-        for (int i = 0; i < iz; ++i)
-            specData.savg[i] = ssum[i] / m_ihsym;
-
-        if (m_ihsym % 10 == 0) {
-            auto const mode4 = nch[nsmo];
-            auto const nsmo = 4 * std::min(10 * mode4, 150);
-
-            flat1(specData.savg, iz, nsmo, specData.slin);
-
-            if (mode4 >= 2) {
-                std::array<float, JS8_NSMAX> tmp;
-
-                smo(specData.slin, tmp.data(), iz, mode4);
-                smo(tmp.data(), specData.slin, iz, mode4);
-            }
-
-            std::fill(std::begin(specData.slin),
-                      std::begin(specData.slin) + 250, 0.0f);
-
-            auto const ia = static_cast<int>(500.0 / m_df3);
-            auto const ib = static_cast<int>(2700.0 / m_df3);
-            auto const smin = *std::min_element(std::begin(specData.slin) + ia,
-                                                std::begin(specData.slin) + ib);
-            auto const smax = *std::max_element(std::begin(specData.slin),
-                                                std::begin(specData.slin) + iz);
-            auto const scale = (smax > smin) ? 50.0f / (smax - smin) : 0.0f;
-
-            for (auto &val : specData.slin)
-                val = std::max(0.0f, scale * (val - smin));
-        }
-    } else if (k < 2048)
-        m_ihsym = 0;
-
-    // make sure ja is equal to k so if we jump ahead in the buffer, everything
-    // resolves correctly
-    ja = k;
-
-    if (m_ihsym <= 0)
-        return;
-
-    if (ui)
-        ui->signal_meter_widget->setValue(m_px, m_pxmax); // Update thermometer
-
-    if (m_monitoring)
-        m_wideGraph->dataSink(s, m_df3);
-
-    decode(k);
-}
+void dataSink(); // JS8_Mainwindow/dataSink.cpp
 
 void MainWindow::showSoundInError(const QString &errorMsg) {
     JS8MessageBox::critical_message(this, tr("Error in Sound Input"), errorMsg);
@@ -4249,550 +3416,7 @@ QDateTime MainWindow::nextTransmitCycle() {
     return timestamp;
 }
 
-QList<int> generateOffsets(int minOffset, int maxOffset) {
-    QList<int> offsets;
-
-    // TODO: these offsets aren't ordered correctly...
-
-    for (int i = minOffset; i <= maxOffset; i++) {
-        offsets.append(i);
-    }
-    return offsets;
-}
-
-void MainWindow::processDecodeEvent(JS8::Event::Variant const &event) {
-    static QList<qint32> driftQueue;
-    static qint32 syncStart = -1;
-
-    std::visit(
-        [this](auto &&e) {
-            using T = std::decay_t<decltype(e)>;
-
-            if constexpr (std::is_same_v<T, JS8::Event::DecodeStarted>) {
-                if (m_wideGraph->shouldDisplayDecodeAttempts()) {
-                    m_wideGraph->drawHorizontalLine(QColor(Qt::yellow), 0, 5);
-                }
-            } else if constexpr (std::is_same_v<T, JS8::Event::SyncStart>) {
-                syncStart = e.position;
-            } else if constexpr (std::is_same_v<T, JS8::Event::SyncState>) {
-                if (m_wideGraph->shouldDisplayDecodeAttempts()) {
-                    auto const drawDecodeLine =
-                        [this, freq = static_cast<int>(e.frequency),
-                         mode = e.mode](QColor const &color) {
-                            m_wideGraph->drawDecodeLine(
-                                color, freq,
-                                freq + JS8::Submode::bandwidth(mode));
-                        };
-
-                    if (e.type == JS8::Event::SyncState::Type::DECODED) {
-                        drawDecodeLine(Qt::red);
-                    } else if (auto const xdtMs = static_cast<int>(e.dt * 1000);
-                               std::abs(xdtMs) <= 2000) {
-                        if (e.sync.candidate < 10)
-                            drawDecodeLine(Qt::darkCyan);
-                        else if (e.sync.candidate <= 15)
-                            drawDecodeLine(Qt::cyan);
-                        else if (e.sync.candidate <= 21)
-                            drawDecodeLine(Qt::white);
-                    }
-                }
-            } else if constexpr (std::is_same_v<T,
-                                                JS8::Event::DecodeFinished>) {
-                qCDebug(decoder_js8) << "decode duration"
-                                     << m_decoderBusyStartTime.msecsTo(
-                                            QDateTime::currentDateTimeUtc())
-                                     << "ms";
-
-                // TODO: move this into a function
-                if (!driftQueue.isEmpty()) {
-                    if (m_driftMsMMA_N == 0) {
-                        m_driftMsMMA_N = 1;
-                        m_driftMsMMA = DriftingDateTime::drift();
-                    }
-
-                    // let the widegraph know for timing control
-                    m_wideGraph->notifyDriftedSignalsDecoded(
-                        driftQueue.count());
-
-                    while (!driftQueue.isEmpty()) {
-                        qint32 newDrift = driftQueue.first();
-                        driftQueue.removeFirst();
-
-                        m_driftMsMMA =
-                            (((m_driftMsMMA_N - 1) * m_driftMsMMA) + newDrift) /
-                            m_driftMsMMA_N;
-                        if (m_driftMsMMA_N < 60)
-                            m_driftMsMMA_N++; // cap it to 60 observations
-                    }
-
-                // XXX The following lines do nothing; it's a completely dead
-                // store. For
-                //     now, just #ifdefing them out, but they were in
-                //     the 2.2.1-devel code, and presumably they were important;
-                //     need to see what the intent was here.
-#if 0
-          qint32 driftLimitMs = JS8::Submode::period(Varicode::JS8CallNormal) * 1000;
-          qint32 newDriftMs   = m_driftMsMMA;
-          if(newDriftMs < 0){
-              newDriftMs = -((-newDriftMs) % driftLimitMs);
-          } else {
-              newDriftMs = ((newDriftMs) % driftLimitMs);
-          }
-#endif
-
-                    setDrift(m_driftMsMMA);
-                    // writeNoticeTextToUI(QDateTime::currentDateTimeUtc(),
-                    // QString("Automatic Drift: %1").arg(driftAvg));
-                }
-
-                m_bDecoded = e.decoded > 0;
-                decodeDone();
-            } else if constexpr (std::is_same_v<T, JS8::Event::Decoded>) {
-                // A frame is valid if we haven't seen the same frame in the
-                // past 1/2 decode period.
-                //
-                // Note: Success here depends on decodes ordered such that
-                // frequencies
-                //       near `dec_data.params.nfqso` arrive here first, so it's
-                //       key to process the decode candidates in an ordered
-                //       manner, likely by sorting the raw take from the initial
-                //       selection pass.
-
-                DecodedText decodedtext(e);
-                FrameCacheKey dedupeKey(decodedtext.submode(),
-                                        decodedtext.frame());
-
-                if (auto const it = m_messageDupeCache.find(dedupeKey);
-                    it != m_messageDupeCache.end()) {
-                    if (it->second.secsTo(QDateTime::currentDateTimeUtc()) <
-                        0.5 * JS8::Submode::period(decodedtext.submode())) {
-                        qCDebug(mainwindow_js8)
-                            << "duplicate frame at" << it->second << "using key"
-                            << QString("%1:%2")
-                                   .arg(dedupeKey.submode)
-                                   .arg(dedupeKey.frame);
-                        return;
-                    }
-                }
-#if 0
-        // frames are valid if they meet our minimum rx threshold for the submode
-        bool bValidFrame = decodedtext.snr() >= JS8::Submode::rxSNRThreshold(decodedtext.submode());
-
-        qCDebug(mainwindow_js8) << "valid" << bValidFrame << JS8::Submode::name(decodedtext.submode()) << "decoded text" << decodedtext.message();
-
-        // skip if invalid
-        if(!bValidFrame) {
-            return;
-        }
-#else
-                qCDebug(mainwindow_js8)
-                    << JS8::Submode::name(decodedtext.submode())
-                    << "decoded text" << decodedtext.message();
-#endif
-                // TODO: move this into a function
-                // compute time drift for non-dupe messages
-                if (m_wideGraph->shouldAutoSyncSubmode(decodedtext.submode())) {
-                    int m = decodedtext.submode();
-                    float xdt = decodedtext.dt();
-
-                    // if we're here at this point, we _should_ be operating a
-                    // decode every second
-                    //
-                    // so we need to figure out where:
-                    //
-                    //   1) this current decode started
-                    //   2) when that cycle _should_ have started
-                    //   3) compute the delta
-                    //   4) apply the drift
-
-                    qint32 periodMs = 1000 * JS8::Submode::period(m);
-
-                    // writeNoticeTextToUI(now, QString("Decode at %1 (kin: %2,
-                    // lastDecoded:
-                    // %3)").arg(syncStart).arg(dec_data.params.kin).arg(m_lastDecodeStartMap.value(m)));
-
-                    float expectedStartDelay =
-                        JS8::Submode::startDelayMS(m) / 1000.0;
-
-                    float decodedSignalTime =
-                        (float)syncStart / (float)JS8_RX_SAMPLE_RATE;
-
-                    // writeNoticeTextToUI(now, QString("--> started at %1
-                    // seconds into the start of my drifted
-                    // minute").arg(decodedSignalTime));
-
-                    // writeNoticeTextToUI(now, QString("--> we add a time delta
-                    // of %1 seconds into the start of the cycle").arg(xdt));
-
-                    // adjust for expected start delay
-                    decodedSignalTime -= expectedStartDelay;
-
-                    // adjust for time delta
-                    decodedSignalTime += xdt;
-
-                    // ensure that we are within a 60 second minute
-                    if (decodedSignalTime < 0) {
-                        decodedSignalTime += 60.0f;
-                    } else if (decodedSignalTime > 60) {
-                        decodedSignalTime -= 60.0f;
-                    }
-
-                    // writeNoticeTextToUI(now, QString("--> so signal adjusted
-                    // started at %1 seconds into the start of my drifted
-                    // minute").arg(decodedSignalTime));
-
-                    qint32 decodedSignalTimeMs = 1000 * decodedSignalTime;
-                    qint32 cycleStartTimeMs =
-                        (decodedSignalTimeMs / periodMs) * periodMs;
-                    qint32 driftMs = cycleStartTimeMs - decodedSignalTimeMs;
-
-                    // writeNoticeTextToUI(now, QString("--> which is a drift
-                    // adjustment of %1 milliseconds").arg(driftMs));
-
-                    // if we have a large negative offset (say -14000), use the
-                    // positive inverse of +1000
-                    if (driftMs + periodMs < qAbs(driftMs)) {
-                        driftMs += periodMs;
-                    }
-                    // if we have a large positive offset (say 14000, use the
-                    // negative inverse of -1000)
-                    else if (qAbs(driftMs - periodMs) < driftMs) {
-                        driftMs -= periodMs;
-                    }
-
-                    // writeNoticeTextToUI(now, QString("--> which is a
-                    // corrected drift adjustment of %1
-                    // milliseconds").arg(driftMs));
-
-                    qint32 newDrift = DriftingDateTime::drift() + driftMs;
-                    if (newDrift < 0) {
-                        newDrift %= -periodMs;
-                    } else {
-                        newDrift %= periodMs;
-                    }
-
-                    // writeNoticeTextToUI(now, QString("--> which is rounded to
-                    // a total drift of %1 milliseconds for this
-                    // period").arg(newDrift));
-
-                    driftQueue.append(newDrift);
-                }
-
-                // if the frame is valid, cache it!
-                m_messageDupeCache.insert_or_assign(
-                    dedupeKey, QDateTime::currentDateTimeUtc());
-
-                // log valid frames to ALL.txt (and correct their timestamp
-                // format)
-                auto freq = dialFrequency();
-
-                // if we changed frequencies, use the old frequency that we
-                // started the decode with
-                if (m_decoderBusyFreq != freq) {
-                    freq = m_decoderBusyFreq;
-                }
-
-                auto date = DriftingDateTime::currentDateTimeUtc().toString(
-                    "yyyy-MM-dd");
-                writeAllTxt(date + " " + decodedtext.string() + " " +
-                            decodedtext.message());
-
-                /**
-                 * @brief Send decode to WSJT-X protocol
-                 *
-                 * Converts JS8Call decode events to WSJT-X Decode messages and
-                 * sends them to WSJT-X protocol clients. For HeartBeat
-                 * messages, ensures the message text includes callsign and grid
-                 * so clients can properly associate them with grid plots.
-                 */
-                if (m_wsjtxMessageMapper && m_config.wsjtx_protocol_enabled()) {
-                    // Convert decode time from JS8Call format to QTime
-                    auto const hms = decode_time(decodedtext.time());
-                    QTime decode_time = QTime(hms.hour, hms.minute, hms.second);
-
-                    // Send decode message
-                    // Use "JS8" as the mode string (WSJT-X expects mode names
-                    // like "FT8", "FT4", "JT9", etc.)
-                    m_wsjtxMessageMapper->sendDecode(
-                        true, // is_new - always true for new decodes
-                        decode_time, decodedtext.snr(), decodedtext.dt(),
-                        static_cast<quint32>(decodedtext.frequencyOffset()),
-                        "JS8", // mode string
-                        decodedtext.message(), decodedtext.isLowConfidence());
-                }
-
-                ActivityDetail d = {};
-                CallDetail cd = {};
-                CommandDetail cmd = {};
-                CallDetail td = {};
-
-            // Parse General Activity
-#if 1
-                bool shouldParseGeneralActivity = true;
-                if (shouldParseGeneralActivity &&
-                    !decodedtext.messageWords().isEmpty()) {
-                    int offset = decodedtext.frequencyOffset();
-
-                    if (!m_bandActivity.contains(offset)) {
-                        int const range =
-                            JS8::Submode::rxThreshold(decodedtext.submode());
-
-                        QList<int> offsets =
-                            generateOffsets(offset - range, offset + range);
-
-                        foreach (int prevOffset, offsets) {
-                            if (!m_bandActivity.contains(prevOffset)) {
-                                continue;
-                            }
-                            m_bandActivity[offset] = m_bandActivity[prevOffset];
-                            m_bandActivity.remove(prevOffset);
-                            break;
-                        }
-                    }
-
-                    // ActivityDetail d = {};
-                    d.isLowConfidence = decodedtext.isLowConfidence();
-                    d.isCompound = decodedtext.isCompound();
-                    d.isDirected = decodedtext.isDirectedMessage();
-                    d.bits = decodedtext.bits();
-                    d.dial = freq;
-                    d.offset = offset;
-                    d.text = decodedtext.message();
-                    d.utcTimestamp = DriftingDateTime::currentDateTimeUtc();
-                    d.snr = decodedtext.snr();
-                    d.isBuffered = false;
-                    d.submode = decodedtext.submode();
-                    d.tdrift = m_wideGraph->shouldAutoSyncSubmode(d.submode)
-                                   ? DriftingDateTime::drift() / 1000.0
-                                   : decodedtext.dt();
-
-                    // if we have any "first" frame, and a buffer is already
-                    // established, clear it...
-                    int prevBufferOffset = -1;
-                    if (((d.bits & Varicode::JS8CallFirst) ==
-                         Varicode::JS8CallFirst) &&
-                        hasExistingMessageBuffer(decodedtext.submode(),
-                                                 d.offset, true,
-                                                 &prevBufferOffset)) {
-                        qCDebug(mainwindow_js8) << "first message encountered, "
-                                                   "clearing existing buffer"
-                                                << prevBufferOffset;
-                        m_messageBuffer.remove(d.offset);
-                    }
-
-                    // if we have a data frame, and a message buffer has been
-                    // established, buffer it...
-                    if (hasExistingMessageBuffer(decodedtext.submode(),
-                                                 d.offset, true,
-                                                 &prevBufferOffset) &&
-                        !decodedtext.isCompound() &&
-                        !decodedtext.isDirectedMessage()) {
-                        qCDebug(mainwindow_js8)
-                            << "buffering data" << d.dial << d.offset << d.text;
-                        d.isBuffered = true;
-                        m_messageBuffer[d.offset].msgs.append(d);
-                        // TODO: incremental display if it's "to" me.
-                    }
-
-                    m_rxActivityQueue.append(d);
-                    m_bandActivity[offset].append(d);
-                    while (m_bandActivity[offset].count() > 10) {
-                        m_bandActivity[offset].removeFirst();
-                    }
-                }
-#endif
-
-            // Process compound callsign commands (put them in cache)"
-#if 1
-                qCDebug(mainwindow_js8) << "decoded" << decodedtext.frameType()
-                                        << decodedtext.isCompound()
-                                        << decodedtext.isDirectedMessage()
-                                        << decodedtext.isHeartbeat();
-                bool shouldProcessCompound = true;
-                if (shouldProcessCompound && decodedtext.isCompound() &&
-                    !decodedtext.isDirectedMessage()) {
-                    cd.call = decodedtext.compoundCall();
-                    cd.grid = decodedtext.extra(); // compound calls via pings
-                                                   // may contain grid...
-                    cd.snr = decodedtext.snr();
-                    cd.dial = freq;
-                    cd.offset = decodedtext.frequencyOffset();
-                    cd.utcTimestamp = DriftingDateTime::currentDateTimeUtc();
-                    cd.bits = decodedtext.bits();
-                    cd.submode = decodedtext.submode();
-                    cd.tdrift = m_wideGraph->shouldAutoSyncSubmode(d.submode)
-                                    ? DriftingDateTime::drift() / 1000.0
-                                    : decodedtext.dt();
-
-                    // Only respond to HEARTBEATS...remember that CQ messages
-                    // are "Alt" pings
-                    if (decodedtext.isHeartbeat()) {
-                        if (decodedtext.isAlt()) {
-                            // this is a cq with a standard or compound call,
-                            // ala "KN4CRD/P: @ALLCALL CQ CQ CQ"
-                            cd.cqTimestamp =
-                                DriftingDateTime::currentDateTimeUtc();
-
-                            // convert CQ to a directed command and process...
-                            cmd.from = cd.call;
-                            cmd.to = "@ALLCALL";
-                            cmd.cmd = " CQ";
-                            cmd.snr = cd.snr;
-                            cmd.bits = cd.bits;
-                            cmd.grid = cd.grid;
-                            cmd.dial = cd.dial;
-                            cmd.offset = cd.offset;
-                            cmd.utcTimestamp = cd.utcTimestamp;
-                            cmd.tdrift = cd.tdrift;
-                            cmd.submode = cd.submode;
-                            cmd.text = decodedtext.message();
-
-                            // TODO: check bits so we only auto respond to
-                            // "finished" cqs
-                            m_rxCommandQueue.append(cmd);
-
-                            // since this is no longer processed here we omit
-                            // logging it here. if we change this behavior, we'd
-                            // change this back to logging here.
-                            // logCallActivity(cd, true);
-
-                            // notification for cq
-                            tryNotify("cq");
-
-                        } else {
-                            // convert HEARTBEAT to a directed command and
-                            // process...
-                            cmd.from = cd.call;
-                            cmd.to = "@HB";
-                            cmd.cmd = " HEARTBEAT";
-                            cmd.snr = cd.snr;
-                            cmd.bits = cd.bits;
-                            cmd.grid = cd.grid;
-                            cmd.dial = cd.dial;
-                            cmd.offset = cd.offset;
-                            cmd.utcTimestamp = cd.utcTimestamp;
-                            cmd.tdrift = cd.tdrift;
-                            cmd.submode = cd.submode;
-
-                            // TODO: check bits so we only auto respond to
-                            // "finished" heartbeats
-                            m_rxCommandQueue.append(cmd);
-
-                            // notification for hb
-                            tryNotify("hb");
-                        }
-
-                    } else {
-                        qCDebug(mainwindow_js8)
-                            << "buffering compound call" << cd.offset << cd.call
-                            << cd.bits;
-
-                        hasExistingMessageBuffer(cd.submode, cd.offset, true,
-                                                 nullptr);
-                        m_messageBuffer[cd.offset].compound.append(cd);
-                    }
-                }
-#endif
-
-            // Parse commands
-            // KN4CRD K1JT ?
-#if 1
-                bool shouldProcessDirected = true;
-                if (shouldProcessDirected && decodedtext.isDirectedMessage()) {
-                    auto parts = decodedtext.directedMessage();
-
-                    cmd.from = parts.at(0);
-                    cmd.to = parts.at(1);
-                    cmd.cmd = parts.at(2);
-                    cmd.dial = freq;
-                    cmd.offset = decodedtext.frequencyOffset();
-                    cmd.snr = decodedtext.snr();
-                    cmd.utcTimestamp = DriftingDateTime::currentDateTimeUtc();
-                    cmd.bits = decodedtext.bits();
-                    cmd.extra =
-                        parts.length() > 2 ? parts.mid(3).join(" ") : "";
-                    cmd.submode = decodedtext.submode();
-                    cmd.tdrift = m_wideGraph->shouldAutoSyncSubmode(cmd.submode)
-                                     ? DriftingDateTime::drift() / 1000.0
-                                     : decodedtext.dt();
-
-                    // if the command is a buffered command and its not the last
-                    // frame OR we have from or to in a separate message
-                    // (compound call)
-                    if ((Varicode::isCommandBuffered(cmd.cmd) &&
-                         (cmd.bits & Varicode::JS8CallLast) !=
-                             Varicode::JS8CallLast) ||
-                        cmd.from == "<....>" || cmd.to == "<....>") {
-                        qCDebug(mainwindow_js8)
-                            << "buffering cmd" << cmd.dial << cmd.offset
-                            << cmd.cmd << cmd.from << cmd.to;
-
-                        // log complete buffered callsigns immediately
-                        if (cmd.from != "<....>" && cmd.to != "<....>") {
-                            CallDetail cmdcd = {};
-                            cmdcd.call = cmd.from;
-                            cmdcd.bits = cmd.bits;
-                            cmdcd.snr = cmd.snr;
-                            cmdcd.dial = cmd.dial;
-                            cmdcd.offset = cmd.offset;
-                            cmdcd.utcTimestamp = cmd.utcTimestamp;
-                            cmdcd.ackTimestamp =
-                                cmd.to == m_config.my_callsign()
-                                    ? cmd.utcTimestamp
-                                    : QDateTime{};
-                            cmdcd.tdrift = cmd.tdrift;
-                            cmdcd.submode = cmd.submode;
-                            logCallActivity(cmdcd, false);
-                            logHeardGraph(cmd.from, cmd.to);
-                        }
-
-                        // merge any existing buffer to this frequency
-                        hasExistingMessageBuffer(cmd.submode, cmd.offset, true,
-                                                 nullptr);
-
-                        if (cmd.to == m_config.my_callsign()) {
-                            d.shouldDisplay = true;
-                        }
-
-                        m_messageBuffer[cmd.offset].cmd = cmd;
-                        m_messageBuffer[cmd.offset].msgs.clear();
-                    } else {
-                        m_rxCommandQueue.append(cmd);
-                    }
-
-                    // check to see if this is a station we've heard 3rd party
-                    bool shouldCaptureThirdPartyCallsigns = false;
-                    if (shouldCaptureThirdPartyCallsigns &&
-                        Radio::base_callsign(cmd.to) !=
-                            Radio::base_callsign(m_config.my_callsign())) {
-                        QString relayCall =
-                            QString("%1|%2")
-                                .arg(Radio::base_callsign(cmd.from))
-                                .arg(Radio::base_callsign(cmd.to));
-                        int snr = -100;
-                        if (parts.length() == 4) {
-                            snr = QString(parts.at(3)).toInt();
-                        }
-
-                        // CallDetail td = {};
-                        td.through = cmd.from;
-                        td.call = cmd.to;
-                        td.grid = "";
-                        td.snr = snr;
-                        td.dial = cmd.dial;
-                        td.offset = cmd.offset;
-                        td.utcTimestamp = cmd.utcTimestamp;
-                        td.tdrift = cmd.tdrift;
-                        td.submode = cmd.submode;
-                        logCallActivity(td, true);
-                        logHeardGraph(cmd.from, cmd.to);
-                    }
-                }
-#endif
-            }
-        },
-        event);
-}
+void processDecodeEvent(); // JS8_Mainwindow/processDecodeEvent.cpp
 
 bool MainWindow::hasExistingMessageBufferToMe(int *const pOffset) {
     for (auto const [offset, buffer] : m_messageBuffer.asKeyValueRange()) {
@@ -7347,461 +5971,7 @@ void MainWindow::buildCallActivitySortByMenu(QMenu *menu) {
                      {"Mode Speed (fastest first)", "-submode"}});
 }
 
-void MainWindow::buildQueryMenu(QMenu *menu, QString call) {
-    bool isAllCall = isAllCallIncluded(call);
-
-    // for now, we're going to omit displaying the call...delete this if we want
-    // the other functionality
-    call = "";
-
-    auto grid = m_config.my_grid();
-
-    bool emptyInfo = m_config.my_info().isEmpty();
-    bool emptyGrid = m_config.my_grid().isEmpty();
-
-    auto callAction = menu->addAction(
-        QString("Send a directed message to selected callsign"));
-    connect(callAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(QString("%1 ").arg(selectedCall), true);
-    });
-
-    menu->addSeparator();
-
-    auto sendReplyAction = menu->addAction(
-        QString("%1 Reply - Send reply message to selected callsign")
-            .arg(call)
-            .trimmed());
-    connect(sendReplyAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        auto message = m_config.reply_message();
-        message = replaceMacros(message, buildMacroValues(), true);
-        addMessageText(QString("%1 %2").arg(selectedCall).arg(message), true);
-    });
-
-    auto sendSNRAction = menu->addAction(
-        QString("%1 SNR - Send a signal report to the selected callsign")
-            .arg(call)
-            .trimmed());
-    sendSNRAction->setEnabled(m_callActivity.contains(callsignSelected()));
-    connect(sendSNRAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        if (!m_callActivity.contains(selectedCall)) {
-            return;
-        }
-
-        auto d = m_callActivity[selectedCall];
-        addMessageText(QString("%1 SNR %2")
-                           .arg(selectedCall)
-                           .arg(Varicode::formatSNR(d.snr)),
-                       true);
-
-        if (m_config.transmit_directed())
-            toggleTx(true);
-    });
-
-    auto infoAction = menu->addAction(
-        QString("%1 INFO - Send my station information").arg(call).trimmed());
-    infoAction->setDisabled(emptyInfo);
-    connect(infoAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(
-            QString("%1 INFO %2").arg(selectedCall).arg(m_config.my_info()),
-            true);
-
-        if (m_config.transmit_directed())
-            toggleTx(true);
-    });
-
-    auto gridAction = menu->addAction(
-        QString("%1 GRID %2 - Send my current station Maidenhead grid locator")
-            .arg(call)
-            .arg(grid)
-            .trimmed());
-    gridAction->setDisabled(emptyGrid);
-    connect(gridAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(
-            QString("%1 GRID %2").arg(selectedCall).arg(m_config.my_grid()),
-            true);
-
-        if (m_config.transmit_directed())
-            toggleTx(true);
-    });
-
-    menu->addSeparator();
-
-    auto snrQueryAction = menu->addAction(
-        QString("%1 SNR? - What is my signal report?").arg(call).trimmed());
-    snrQueryAction->setDisabled(isAllCall);
-    connect(snrQueryAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(QString("%1 SNR?").arg(selectedCall), true);
-
-        if (m_config.transmit_directed())
-            toggleTx(true);
-    });
-
-    auto infoQueryAction =
-        menu->addAction(QString("%1 INFO? - What is your station information?")
-                            .arg(call)
-                            .trimmed());
-    infoQueryAction->setDisabled(isAllCall);
-    connect(infoQueryAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(QString("%1 INFO?").arg(selectedCall), true);
-
-        if (m_config.transmit_directed())
-            toggleTx(true);
-    });
-
-    auto gridQueryAction =
-        menu->addAction(QString("%1 GRID? - What is your current grid locator?")
-                            .arg(call)
-                            .trimmed());
-    gridQueryAction->setDisabled(isAllCall);
-    connect(gridQueryAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(QString("%1 GRID?").arg(selectedCall), true);
-
-        if (m_config.transmit_directed())
-            toggleTx(true);
-    });
-
-    auto stationIdleQueryAction = menu->addAction(
-        QString("%1 STATUS? - What is your station status message?")
-            .arg(call)
-            .trimmed());
-    stationIdleQueryAction->setDisabled(isAllCall);
-    connect(stationIdleQueryAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(QString("%1 STATUS?").arg(selectedCall), true);
-
-        if (m_config.transmit_directed())
-            toggleTx(true);
-    });
-
-    auto heardQueryAction = menu->addAction(
-        QString("%1 HEARING? - What are the stations are you hearing? (Top 4 "
-                "ranked by most recently heard)")
-            .arg(call)
-            .trimmed());
-    heardQueryAction->setDisabled(isAllCall);
-    connect(heardQueryAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(QString("%1 HEARING?").arg(selectedCall), true);
-
-        if (m_config.transmit_directed())
-            toggleTx(true);
-    });
-
-#if 0
-    auto retransmitAction = menu->addAction(QString("%1|[MESSAGE] - Please ACK and retransmit the following message").arg(call).trimmed());
-    retransmitAction->setDisabled(isAllCall);
-    connect(retransmitAction, &QAction::triggered, this, [this](){
-
-        QString selectedCall = callsignSelected();
-        if(selectedCall.isEmpty()){
-            return;
-        }
-
-        addMessageText(QString("%1|[MESSAGE]").arg(selectedCall), true, true);
-    });
-#endif
-
-    auto alertAction = menu->addAction(
-        QString("%1>[MESSAGE] - Please relay this message to its destination")
-            .arg(call)
-            .trimmed());
-    alertAction->setDisabled(isAllCall);
-    connect(alertAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(QString("%1>[MESSAGE]").arg(selectedCall), true, true);
-    });
-
-    auto msgAction = menu->addAction(
-        QString("%1 MSG [MESSAGE] - Please store this message in your inbox")
-            .arg(call)
-            .trimmed());
-    msgAction->setDisabled(isAllCall);
-    connect(msgAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(QString("%1 MSG [MESSAGE]").arg(selectedCall), true,
-                       true);
-    });
-
-    auto msgToAction = menu->addAction(
-        QString("%1 MSG TO:[CALLSIGN] [MESSAGE] - Please store this message at "
-                "your station for later retreival by [CALLSIGN]")
-            .arg(call)
-            .trimmed());
-    msgToAction->setDisabled(isAllCall);
-    connect(msgToAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(
-            QString("%1 MSG TO:[CALLSIGN] [MESSAGE]").arg(selectedCall), true,
-            true);
-    });
-
-    auto qsoQueryAction = menu->addAction(
-        QString("%1 QUERY CALL [CALLSIGN]? - Please acknowledge you can "
-                "communicate directly with [CALLSIGN]")
-            .arg(call)
-            .trimmed());
-    connect(qsoQueryAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(QString("%1 QUERY CALL [CALLSIGN]?").arg(selectedCall),
-                       true, true);
-    });
-
-    auto qsoQueryMsgsAction = menu->addAction(
-        QString("%1 QUERY MSGS - Do you have any messages for me?")
-            .arg(call)
-            .trimmed());
-    connect(qsoQueryMsgsAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(QString("%1 QUERY MSGS").arg(selectedCall), true, true);
-    });
-
-    auto qsoQueryMsgAction =
-        menu->addAction(QString("%1 QUERY MSG [ID] - Please deliver the "
-                                "complete message identified by ID")
-                            .arg(call)
-                            .trimmed());
-    connect(qsoQueryMsgAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(QString("%1 QUERY MSG [ID]").arg(selectedCall), true,
-                       true);
-    });
-
-    menu->addSeparator();
-
-    auto agnAction = menu->addAction(
-        QString("%1 AGN? - Please repeat your last transmission")
-            .arg(call)
-            .trimmed());
-    connect(agnAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(QString("%1 AGN?").arg(selectedCall), true);
-
-        if (m_config.transmit_directed())
-            toggleTx(true);
-    });
-
-    auto qslQueryAction = menu->addAction(
-        QString("%1 QSL? - Did you receive my last transmission?")
-            .arg(call)
-            .trimmed());
-    connect(qslQueryAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(QString("%1 QSL?").arg(selectedCall), true);
-
-        if (m_config.transmit_directed())
-            toggleTx(true);
-    });
-
-    auto qslAction = menu->addAction(
-        QString("%1 QSL - I confirm I received your last transmission")
-            .arg(call)
-            .trimmed());
-    connect(qslAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(QString("%1 QSL").arg(selectedCall), true);
-
-        if (m_config.transmit_directed())
-            toggleTx(true);
-    });
-
-    auto yesAction = menu->addAction(
-        QString("%1 YES - I confirm your last inquiry").arg(call).trimmed());
-    connect(yesAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(QString("%1 YES").arg(selectedCall), true);
-
-        if (m_config.transmit_directed())
-            toggleTx(true);
-    });
-
-    auto noAction =
-        menu->addAction(QString("%1 NO - I do not confirm your last inquiry")
-                            .arg(call)
-                            .trimmed());
-    connect(noAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(QString("%1 NO").arg(selectedCall), true);
-
-        if (m_config.transmit_directed())
-            toggleTx(true);
-    });
-
-    auto hwAction = menu->addAction(
-        QString("%1 HW CPY? - How do you copy?").arg(call).trimmed());
-    connect(hwAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(QString("%1 HW CPY?").arg(selectedCall), true);
-
-        if (m_config.transmit_directed())
-            toggleTx(true);
-    });
-
-    auto rrAction = menu->addAction(
-        QString("%1 RR - Roger. Received. I copy.").arg(call).trimmed());
-    connect(rrAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(QString("%1 RR").arg(selectedCall), true);
-
-        if (m_config.transmit_directed())
-            toggleTx(true);
-    });
-
-    auto fbAction =
-        menu->addAction(QString("%1 FB - Fine Business").arg(call).trimmed());
-    connect(fbAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(QString("%1 FB").arg(selectedCall), true);
-
-        if (m_config.transmit_directed())
-            toggleTx(true);
-    });
-
-    auto sevenThreeAction = menu->addAction(
-        QString("%1 73 - I send my best regards").arg(call).trimmed());
-    connect(sevenThreeAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(QString("%1 73").arg(selectedCall), true);
-
-        if (m_config.transmit_directed())
-            toggleTx(true);
-    });
-
-    auto skAction =
-        menu->addAction(QString("%1 SK - End of contact").arg(call).trimmed());
-    connect(skAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(QString("%1 SK").arg(selectedCall), true);
-
-        if (m_config.transmit_directed())
-            toggleTx(true);
-    });
-
-    auto ditDitAction = menu->addAction(
-        QString("%1 DIT DIT - End of contact / Two bits").arg(call).trimmed());
-    connect(ditDitAction, &QAction::triggered, this, [this]() {
-        QString selectedCall = callsignSelected();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        addMessageText(QString("%1 DIT DIT").arg(selectedCall), true);
-
-        if (m_config.transmit_directed())
-            toggleTx(true);
-    });
-}
+void buildQueryMenu(); // JS8_Mainwindow/buildQueryMenu.cpp
 
 void MainWindow::buildRelayMenu(QMenu *menu) {
     auto now = DriftingDateTime::currentDateTimeUtc();
@@ -9144,176 +7314,7 @@ void MainWindow::processIdleActivity() {
     }
 }
 
-void MainWindow::processRxActivity() {
-    if (m_rxActivityQueue.isEmpty()) {
-        return;
-    }
-
-    int freqOffset = freq();
-
-    qCDebug(mainwindow_js8)
-        << m_messageBuffer.count() << "message buffers open";
-
-    while (!m_rxActivityQueue.isEmpty()) {
-        ActivityDetail d = m_rxActivityQueue.dequeue();
-
-        if (canSendNetworkMessage()) {
-            sendNetworkMessage(
-                "RX.ACTIVITY", d.text,
-                {{"_ID", QVariant(-1)},
-                 {"FREQ", QVariant(d.dial + d.offset)},
-                 {"DIAL", QVariant(d.dial)},
-                 {"OFFSET", QVariant(d.offset)},
-                 {"SNR", QVariant(d.snr)},
-                 {"SPEED", QVariant(d.submode)},
-                 {"TDRIFT", QVariant(d.tdrift)},
-                 {"UTC", QVariant(d.utcTimestamp.toMSecsSinceEpoch())}});
-        }
-
-        // use the actual frequency and check its delta from our current
-        // frequency meaning, if our current offset is 1502 and the d.freq is
-        // 1492, the delta is <= 10;
-        bool shouldDisplay =
-            abs(d.offset - freqOffset) <= JS8::Submode::rxThreshold(d.submode);
-
-        int prevOffset = d.offset;
-        if (hasExistingMessageBuffer(d.submode, d.offset, false, &prevOffset) &&
-            ((m_messageBuffer[prevOffset].cmd.to == m_config.my_callsign()) ||
-             // (isAllCallIncluded(m_messageBuffer[prevOffset].cmd.to))     ||
-             // // uncomment this if we want to incrementally print allcalls
-             (isGroupCallIncluded(m_messageBuffer[prevOffset].cmd.to)))) {
-            d.isBuffered = true;
-            shouldDisplay = true;
-
-            if (!m_messageBuffer[prevOffset].compound.isEmpty()) {
-                // qCDebug(mainwindow_js8) << "should display compound too
-                // because at this point it hasn't been displayed" <<
-                // m_messageBuffer[prevOffset].compound.last().call;
-
-                auto lastCompound = m_messageBuffer[prevOffset].compound.last();
-
-                // fixup compound call incremental text
-                d.text = QString("%1: %2").arg(lastCompound.call).arg(d.text);
-                d.utcTimestamp =
-                    qMin(d.utcTimestamp, lastCompound.utcTimestamp);
-            }
-
-        } else if (hasClosedExistingMessageBuffer(d.offset)) {
-            // incremental typeahead should just be displayed...
-            // TODO: should the buffer be reopened?
-            shouldDisplay = true;
-
-        } else if (d.isDirected && d.text.contains("<....>")) {
-            // if this is a _partial_ directed message, skip until the complete
-            // call comes through.
-            continue;
-
-        } else if (d.isDirected &&
-                   (d.text.contains(": HB ") ||
-                    d.text.contains(": @ALLCALL HB"))) { // TODO: HEARTBEAT
-            // if this is a heartbeat, process elsewhere...
-            continue;
-        }
-
-        // if this is the first data frame of a standard message, parse the
-        // first word callsigns and spot them :)
-        if ((d.bits & Varicode::JS8CallFirst) == Varicode::JS8CallFirst &&
-            !d.isDirected && !d.isCompound) {
-            auto calls = Varicode::parseCallsigns(d.text);
-            if (!calls.isEmpty()) {
-                auto theirCall = calls.first();
-                if (d.text.startsWith(theirCall) &&
-                    d.text.mid(theirCall.length(), 1) == ":") {
-                    CallDetail cd = {};
-                    cd.call = theirCall;
-                    cd.dial = d.dial;
-                    cd.offset = d.offset;
-                    cd.snr = d.snr;
-                    cd.bits = d.bits;
-                    cd.tdrift = d.tdrift;
-                    cd.utcTimestamp = d.utcTimestamp;
-                    cd.submode = d.submode;
-                    logCallActivity(cd, true);
-                }
-            }
-        }
-
-        // TODO: incremental printing of directed messages
-        // Display if:
-        // 1) this is a directed message header "to" us and should be
-        // buffered... 2) or, this is a buffered message frame for a buffer with
-        // us as the recipient.
-
-        if (!shouldDisplay) {
-            continue;
-        }
-
-        bool isFirst =
-            (d.bits & Varicode::JS8CallFirst) == Varicode::JS8CallFirst;
-        bool isLast = (d.bits & Varicode::JS8CallLast) == Varicode::JS8CallLast;
-
-        // if we're the last message, let's display our EOT character
-        if (isLast) {
-            d.text = QString("%1 %2 ")
-                         .arg(Varicode::rstrip(d.text))
-                         .arg(m_config.eot());
-        }
-
-        // log it to the display!
-        displayTextForFreq(d.text, d.offset, d.utcTimestamp, false, isFirst,
-                           isLast);
-
-        // If we've received a message to be displayed, we should no longer call
-        // CQ.
-        if (m_cq_loop->isActive()) {
-            qCDebug(mainwindow_js8)
-                << "Canceling calling CQ loop to priorize incoming messages.";
-            m_cq_loop->onLoopCancel();
-        }
-
-        if (isLast) {
-            clearOffsetDirected(d.offset);
-        }
-
-        if (isLast && !d.isBuffered) {
-            // buffered commands need the rxFrameBlockNumbers cache so it can
-            // fixup its display all other "last" data frames can clear the
-            // rxFrameBlockNumbers cache so the next message will be on a new
-            // line.
-            m_rxFrameBlockNumbers.remove(d.offset);
-        }
-    }
-
-#if 0
-    // TODO: this works but should also print in the rx window.
-    foreach(auto offset, m_bandActivity.keys()){
-        if(seen.contains(offset)){
-            continue;
-        }
-
-        if(m_bandActivity[offset].isEmpty()){
-            continue;
-        }
-
-        auto last = m_bandActivity[offset].last();
-        if((last.bits & Varicode::JS8CallLast) == Varicode::JS8CallLast){
-            continue;
-        }
-
-        auto now = DriftingDateTime::currentDateTimeUtc();
-        if(last.utcTimestamp.secsTo(now) < m_TRperiod){
-            continue;
-        }
-
-        ActivityDetail d = {};
-        d.text = " . . . ";
-        d.utcTimestamp = now;
-        d.snr = -99;
-
-        m_bandActivity[offset].append(d);
-    }
-#endif
-}
+void processRxActivity(); // JS8_Mainwindow/processRxActivity.cpp
 
 void MainWindow::processCompoundActivity() {
     if (m_messageBuffer.isEmpty()) {
@@ -9509,956 +7510,7 @@ void MainWindow::processBufferedActivity() {
     }
 }
 
-void MainWindow::processCommandActivity() {
-#if 0
-    if (!m_txFrameQueue.isEmpty()) {
-        return;
-    }
-#endif
-
-    if (m_rxCommandQueue.isEmpty()) {
-        return;
-    }
-
-#if 0
-    bool processed = false;
-
-    int f = currentFreq();
-#endif
-
-    auto now = DriftingDateTime::currentDateTimeUtc();
-
-    while (!m_rxCommandQueue.isEmpty()) {
-        auto d = m_rxCommandQueue.dequeue();
-
-        auto selectedCallsign = callsignSelected();
-        bool isAllCall = isAllCallIncluded(d.to);
-        bool isGroupCall = isGroupCallIncluded(d.to);
-
-        qCDebug(mainwindow_js8)
-            << "try processing command" << d.from << d.to << d.cmd << d.dial
-            << d.offset << d.grid << d.extra << isAllCall << isGroupCall;
-
-        // if we need a compound callsign but never got one...skip
-        if (d.from == "<....>" || d.to == "<....>") {
-            continue;
-        }
-
-        // we're only processing a subset of queries at this point
-        if (!Varicode::isCommandAllowed(d.cmd)) {
-            continue;
-        }
-
-        // is this to me?
-        bool toMe =
-            d.to == m_config.my_callsign().trimmed() ||
-            d.to == Radio::base_callsign(m_config.my_callsign()).trimmed();
-
-        // log call activity...
-        CallDetail cd = {};
-        cd.call = d.from;
-        cd.grid = d.grid;
-        cd.snr = d.snr;
-        cd.dial = d.dial;
-        cd.offset = d.offset;
-        cd.bits = d.bits;
-        cd.ackTimestamp =
-            d.text.contains(": ACK") || toMe ? d.utcTimestamp : QDateTime{};
-        cd.utcTimestamp = d.utcTimestamp;
-        cd.tdrift = d.tdrift;
-        cd.submode = d.submode;
-        logCallActivity(cd, true);
-        logHeardGraph(d.from, d.to);
-
-        // Placeholder for callbacks to be passed to confirmThenEnqueueMessage
-        // and enqueueMessage
-        Callback callback = nullptr;
-
-        // PROCESS BUFFERED HEARING FOR EVERYONE
-        if (d.cmd == " HEARING") {
-            // 1. parse callsigns
-            // 2. log it to the heard graph
-            auto calls = Varicode::parseCallsigns(d.text);
-            foreach (auto call, calls) {
-                logHeardGraph(d.from, call);
-            }
-        }
-
-        // PROCESS BUFFERED GRID FOR EVERYONE
-        if (d.cmd == " GRID") {
-            // 1. parse grids
-            // 2. log it to our call activity
-            auto grids = Varicode::parseGrids(d.text);
-            foreach (auto grid, grids) {
-                CallDetail cd = {};
-                cd.bits = d.bits;
-                cd.call = d.from;
-                cd.dial = d.dial;
-                cd.offset = d.offset;
-                cd.grid = grid;
-                cd.snr = d.snr;
-                cd.utcTimestamp = d.utcTimestamp;
-                cd.tdrift = d.tdrift;
-                cd.submode = d.submode;
-
-                // PROCESS GRID SPOTS TO APRSIS FOR EVERYONE
-                if (d.to == "@APRSIS") {
-                    spotAprsGrid(cd.dial, cd.offset, cd.snr, cd.call, cd.grid);
-                }
-
-                logCallActivity(cd, true);
-            }
-        }
-
-        // PROCESS @JS8NET, @APRSIS, AND OTHER GROUP SPOTS FOR EVERYONE
-        if (d.to.startsWith("@")) {
-            spotCmd(d);
-        }
-
-        // PROCESS @APRSIS CMD SPOTS FOR EVERYONE
-        if (d.to == "@APRSIS") {
-            spotAprsCmd(d);
-        }
-
-        // PREPARE CMD TEXT STRING
-        QStringList textList = {
-            QString("%1: %2%3").arg(d.from).arg(d.to).arg(d.cmd)};
-
-        if (!d.extra.isEmpty()) {
-            textList.append(d.extra);
-        }
-
-        if (!d.text.isEmpty()) {
-            textList.append(d.text);
-        }
-
-        QString text = textList.join(" ");
-        bool isLast = (d.bits & Varicode::JS8CallLast) == Varicode::JS8CallLast;
-        if (isLast) {
-            // append the eot character to the text
-            text = QString("%1 %2 ")
-                       .arg(Varicode::rstrip(text))
-                       .arg(m_config.eot());
-        }
-
-        // log the text to directed txt log
-        writeMsgTxt(text, d.snr, d.offset);
-
-        // write all directed messages to api
-        if (canSendNetworkMessage()) {
-            sendNetworkMessage(
-                "RX.DIRECTED", text,
-                {{"_ID", QVariant(-1)},
-                 {"FROM", QVariant(d.from)},
-                 {"TO", QVariant(d.to)},
-                 {"CMD", QVariant(d.cmd)},
-                 {"GRID", QVariant(d.grid)},
-                 {"EXTRA", QVariant(d.extra)},
-                 {"TEXT", QVariant(text)},
-                 {"FREQ", QVariant(d.dial + d.offset)},
-                 {"DIAL", QVariant(d.dial)},
-                 {"OFFSET", QVariant(d.offset)},
-                 {"SNR", QVariant(d.snr)},
-                 {"SPEED", QVariant(d.submode)},
-                 {"TDRIFT", QVariant(d.tdrift)},
-                 {"UTC", QVariant(d.utcTimestamp.toMSecsSinceEpoch())}});
-        }
-
-        // we're only responding to allcalls if we are participating in the
-        // allcall group but, don't avoid for heartbeats...those are technically
-        // allcalls but are processed differently
-        if (isAllCall && m_config.avoid_allcall() && d.cmd != " CQ" &&
-            d.cmd != " HB" && d.cmd != " HEARTBEAT") {
-            continue;
-        }
-
-        // we're only responding to allcall, groupcalls, and our callsign at
-        // this point, so we'll end after logging the callsigns we've heard
-        if (!isAllCall && !toMe && !isGroupCall) {
-            continue;
-        }
-
-        ActivityDetail ad = {};
-        ad.isLowConfidence = false;
-        ad.isDirected = true;
-        ad.bits = d.bits;
-        ad.dial = d.dial;
-        ad.offset = d.offset;
-        ad.snr = d.snr;
-        ad.text = text;
-        ad.utcTimestamp = d.utcTimestamp;
-
-        // we'd be double printing here if were on frequency, so let's be
-        // "smart" about this...
-        bool shouldDisplay = true;
-
-        // don't display ping allcalls
-        if (isAllCall && (d.cmd != " " || ad.text.contains("@HB HEARTBEAT"))) {
-            shouldDisplay = false;
-        }
-
-        if (shouldDisplay) {
-            auto c = ui->textEditRX->textCursor();
-            c.movePosition(QTextCursor::End);
-            ui->textEditRX->setTextCursor(c);
-
-            // ACKs and SNRs are the most likely source of items to be
-            // overwritten (multiple responses at once)... so don't overwrite
-            // those (i.e., print each on a new line)
-            bool shouldOverwrite =
-                (!d.cmd.contains(" ACK") &&
-                 !d.cmd.contains(" SNR")); /* && isRecentOffset(d.freq);*/
-
-            if (shouldOverwrite &&
-                ui->textEditRX->find(d.utcTimestamp.time().toString(),
-                                     QTextDocument::FindBackward)) {
-                // ... maybe we could delete the last line that had this message
-                // on this frequency...
-                c = ui->textEditRX->textCursor();
-                c.movePosition(QTextCursor::StartOfBlock);
-                c.movePosition(QTextCursor::EndOfBlock,
-                               QTextCursor::KeepAnchor);
-                qCDebug(mainwindow_js8) << "should display directed message, "
-                                           "erasing last rx activity line..."
-                                        << c.selectedText().toUpper();
-                c.removeSelectedText();
-                c.deletePreviousChar();
-                c.deletePreviousChar();
-                /*
-                c.deleteChar();
-                c.deleteChar();
-                */
-            }
-
-            // log it to the display!
-            displayTextForFreq(ad.text, ad.offset, ad.utcTimestamp, false, true,
-                               false);
-
-            /*
-            // and send it to the network in case we want to interact with it
-            from an external app... if(canSendNetworkMessage()){
-                sendNetworkMessage("RX.DIRECTED.ME", ad.text, {
-                    {"_ID", QVariant(-1)},
-                    {"FROM", QVariant(d.from)},
-                    {"TO", QVariant(d.to)},
-                    {"CMD", QVariant(d.cmd)},
-                    {"GRID", QVariant(d.grid)},
-                    {"EXTRA", QVariant(d.extra)},
-                    {"TEXT", QVariant(d.text)},
-                    {"FREQ", QVariant(ad.dial+ad.offset)},
-                    {"DIAL", QVariant(ad.dial)},
-                    {"OFFSET", QVariant(ad.offset)},
-                    {"SNR", QVariant(ad.snr)},
-                    {"SPEED", QVariant(ad.submode)},
-                    {"TDRIFT", QVariant(ad.tdrift)},
-                    {"UTC", QVariant(ad.utcTimestamp.toMSecsSinceEpoch())}
-                });
-            }
-            */
-
-            if (!isAllCall) {
-                // If we've received a message to be displayed,
-                // we should no longer call CQ
-                if (m_cq_loop->isActive()) {
-                    qCDebug(mainwindow_js8)
-                        << "Canceling CQ loop to prioritize incoming messages, "
-                           "case II";
-                    m_cq_loop->onLoopCancel();
-                }
-
-                // notification for directed message
-                tryNotify("directed");
-            }
-        }
-
-        // we're only responding to callsigns in our whitelist if we have one
-        // defined... make sure the whitelist is empty (no restrictions) or the
-        // from callsign or its base callsign is on it
-        auto whitelist = m_config.auto_whitelist();
-        if (!whitelist.isEmpty() &&
-            !(whitelist.contains(d.from) ||
-              whitelist.contains(Radio::base_callsign(d.from)))) {
-            qCDebug(mainwindow_js8)
-                << "skipping command for whitelist" << d.from;
-            continue;
-        }
-
-        // we'll never reply to a blacklisted callsign or base callsign
-        auto blacklist = m_config.auto_blacklist();
-        if (!blacklist.isEmpty() &&
-            (blacklist.contains(d.from) ||
-             blacklist.contains(Radio::base_callsign(d.from)))) {
-            qCDebug(mainwindow_js8)
-                << "skipping command for blacklist" << d.from;
-            continue;
-        }
-
-        // if this is an allcall, check to make sure we haven't replied to their
-        // allcall recently (in the past ten minutes) that way we never get
-        // spammed by allcalls at too high of a frequency
-        if (isAllCall && m_txAllcallCommandCache.contains(d.from) &&
-            m_txAllcallCommandCache[d.from]->secsTo(now) / 60 < 15) {
-            qCDebug(mainwindow_js8)
-                << "skipping command for allcall timeout" << d.from;
-            continue;
-        }
-
-        // don't actually process any automatic message replies while in idle
-        if (m_tx_watchdog) {
-            qCDebug(mainwindow_js8)
-                << "skipping command for idle timeout" << d.from;
-            continue;
-        }
-
-        // HACK: if this is an autoreply cmd and relay path is populated and cmd
-        // is not MSG or MSG TO:, then swap out the relay path
-        if (Varicode::isCommandAutoreply(d.cmd) && !d.relayPath.isEmpty() &&
-            !d.cmd.startsWith(" MSG") && !d.cmd.startsWith(" QUERY")) {
-            d.from = d.relayPath;
-        }
-
-        // construct a reply, if needed
-        QString reply;
-        int priority = PriorityNormal;
-        int freq = -1;
-
-        // QUERIED SNR
-        // Only reply to a SNR request if there is no incoming MSG's and is not
-        // directed to AllCall
-        if (d.cmd == " SNR?" && !isAllCall && m_messageBuffer.isEmpty()) {
-            reply = QString("%1 SNR %2")
-                        .arg(d.from)
-                        .arg(Varicode::formatSNR(d.snr));
-        }
-
-        // QUERIED INFO
-        // Only reply to a INFO request if there is no incoming MSG's and is not
-        // directed to AllCall
-        else if (d.cmd == " INFO?" && !isAllCall && m_messageBuffer.isEmpty()) {
-            QString info = m_config.my_info();
-            if (info.isEmpty()) {
-                continue;
-            }
-
-            reply = QString("%1 INFO %2")
-                        .arg(d.from)
-                        .arg(replaceMacros(info, buildMacroValues(), true));
-        }
-
-        // QUERIED ACTIVE
-        // Only reply to a STATUS request if there is no incoming MSG's and is
-        // not directed to AllCall
-        else if (d.cmd == " STATUS?" && !isAllCall &&
-                 m_messageBuffer.isEmpty()) {
-            QString status = m_config.my_status();
-            if (status.isEmpty()) {
-                continue;
-            }
-
-            reply = QString("%1 STATUS %2")
-                        .arg(d.from)
-                        .arg(replaceMacros(status, buildMacroValues(), true));
-        }
-
-        // QUERIED GRID
-        // Only reply to a GRID request if there is no incoming MSG's and is not
-        // directed to AllCall
-        else if (d.cmd == " GRID?" && !isAllCall && m_messageBuffer.isEmpty()) {
-            QString grid = m_config.my_grid();
-            if (grid.isEmpty()) {
-                continue;
-            }
-
-            reply = QString("%1 GRID %2").arg(d.from).arg(grid);
-        }
-
-        // QUERIED STATIONS HEARD
-        // Only reply to a HEARING request if there is no incoming MSG's and is
-        // not directed to AllCall
-        else if (d.cmd == " HEARING?" && !isAllCall &&
-                 m_messageBuffer.isEmpty()) {
-            auto calls = m_callActivity.keys();
-
-            std::stable_sort(calls.begin(), calls.end(),
-                             [this](QString const &lhs, QString const &rhs) {
-                                 return m_callActivity[rhs].utcTimestamp <
-                                        m_callActivity[lhs].utcTimestamp;
-                             });
-
-            auto const callsignAging = m_config.callsign_aging();
-            auto const maxStations = 4;
-            auto i = 0;
-            QStringList lines;
-
-            for (auto const &call : calls) {
-                if (i >= maxStations)
-                    break;
-                if (call == d.from)
-                    continue;
-
-                auto const &cd = m_callActivity[call];
-
-                if (callsignAging &&
-                    cd.utcTimestamp.secsTo(now) / 60 >= callsignAging) {
-                    continue;
-                }
-
-                lines.append(cd.call);
-                i++;
-            }
-
-            lines.prepend(QString("%1 HEARING").arg(d.from));
-            reply = lines.join(' ');
-        }
-
-        // PROCESS RELAY
-        else if (d.cmd == ">" && !isAllCall && !m_config.relay_off()) {
-
-            // 1. see if there are any more hops to process
-            // 2. if so, forward
-            // 3. otherwise, display alert & reply dialog
-
-            QString callToPattern = {
-                R"(^(?<callsign>\b(?<prefix>[A-Z0-9]{1,4}\/)?(?<base>([0-9A-Z])?([0-9A-Z])([0-9])([A-Z])?([A-Z])?([A-Z])?)(?<suffix>\/[A-Z0-9]{1,4})?(?<type>[> ]))\b)"};
-            QRegularExpression re(callToPattern);
-            auto text = d.text;
-            auto match = re.match(text);
-
-            // if the text starts with a callsign, and relay is not disabled,
-            // and this is not a group callsign, then relay.
-            if (match.hasMatch() && !isGroupCall) {
-                // replace freetext with relayed free text
-                if (match.captured("type") != ">") {
-                    text = text.replace(match.capturedStart("type"),
-                                        match.capturedLength("type"), ">");
-                }
-                reply = QString("%1 *DE* %2").arg(text).arg(d.from);
-
-                // otherwise, as long as we're not an ACK...alert the user and
-                // either send an ACK or Message
-            } else if (!d.text.startsWith("ACK")) {
-
-                // parse out the callsign path
-                auto calls = parseRelayPathCallsigns(d.from, d.text);
-
-                // put these third party calls in the heard list
-                foreach (auto call, calls) {
-                    CallDetail cd = {};
-                    cd.call = call;
-                    cd.snr = -64;
-                    cd.dial = d.dial;
-                    cd.offset = d.offset;
-                    cd.through = d.from;
-                    cd.utcTimestamp = DriftingDateTime::currentDateTimeUtc();
-                    cd.tdrift = d.tdrift;
-                    cd.submode = d.submode;
-                    logCallActivity(cd, false);
-                }
-
-                d.relayPath = calls.join('>');
-
-                reply = QString("%1 ACK").arg(d.relayPath);
-
-                // check to see if the relay text contains a command that should
-                // be replied to instead of an ack.
-                QStringList relayedCmds = d.text.split(" ");
-                if (!relayedCmds.isEmpty()) {
-                    auto first = relayedCmds.first();
-
-                    auto valid = Varicode::isCommandAllowed(first);
-                    if (!valid) {
-                        first = " " + first;
-                        valid = Varicode::isCommandAllowed(first);
-                        if (valid) {
-                            relayedCmds.removeFirst();
-                        }
-                    }
-
-                    // HACK: "MSG TO:" should be supported but contains a space
-                    // :(
-                    if (!relayedCmds.isEmpty()) {
-                        if (first == " MSG") {
-                            auto second = relayedCmds.first();
-                            if (second == "TO:") {
-                                first = " MSG TO:";
-                                relayedCmds.removeFirst();
-                            } else if (second.startsWith("TO:")) {
-                                first = " MSG TO:";
-                                relayedCmds.replace(0, second.mid(3));
-                            }
-                        } else if (first == " QUERY") {
-                            auto second = relayedCmds.first();
-                            if (second == "MSGS" || second == "MSGS?") {
-                                first = " QUERY MSGS";
-                                relayedCmds.removeFirst();
-                            } else if (second == "CALL") {
-                                first = " QUERY CALL";
-                                relayedCmds.removeFirst();
-                            }
-                        }
-                    }
-
-                    if (Varicode::isCommandAllowed(first) &&
-                        Varicode::isCommandAutoreply(first)) {
-                        CommandDetail rd = {};
-                        rd.bits = d.bits;
-                        rd.cmd = first;
-                        rd.dial = d.dial;
-                        rd.offset = d.offset;
-                        rd.from =
-                            d.from; // note, MSG and QUERY commands are not set
-                                    // with from as the relay path.
-                        rd.relayPath = d.relayPath;
-                        rd.text = relayedCmds.join(" "); // d.text;
-                        rd.to = d.to;
-                        rd.utcTimestamp = d.utcTimestamp;
-
-                        m_rxCommandQueue.insert(0, rd);
-                        continue;
-                    }
-                }
-
-#if STORE_RELAY_MSGS_TO_INBOX
-                // if we make it here, this is a message
-                addCommandToMyInbox(d);
-#endif
-            }
-        }
-
-        // PROCESS MESSAGE STORAGE
-        else if (d.cmd == " MSG TO:" && !isAllCall && !m_config.relay_off()) {
-
-            // store message
-            QStringList segs = d.text.split(" ");
-            if (segs.isEmpty()) {
-                continue;
-            }
-
-            auto to = segs.first();
-            segs.removeFirst();
-
-            auto text = segs.join(" ").trimmed();
-
-            auto calls = parseRelayPathCallsigns(d.from, text);
-            d.relayPath = calls.join(">");
-
-            CommandDetail cd = {};
-            cd.bits = d.bits;
-            cd.cmd = d.cmd;
-            cd.extra = d.extra;
-            cd.dial = d.dial;
-            cd.offset = d.offset;
-            cd.from = d.from;
-            cd.grid = d.grid;
-            cd.relayPath = d.relayPath;
-            cd.snr = d.snr;
-            cd.tdrift = d.tdrift;
-            cd.text = text;
-            cd.to = Radio::base_callsign(to);
-            cd.utcTimestamp = d.utcTimestamp;
-            cd.submode = d.submode;
-
-            qCDebug(mainwindow_js8)
-                << "storing message to" << to << ":" << text;
-
-            addCommandToStorage("STORE", cd);
-
-            // we haven't replaced the from with the relay path, so we have to
-            // use it for the ack if there is one
-            reply = QString("%1 ACK").arg(calls.length() > 1 ? d.relayPath
-                                                             : d.from);
-        }
-
-        // PROCESS AGN
-        else if (d.cmd == " AGN?" && !isAllCall && !isGroupCall &&
-                 !m_lastTxMessage.isEmpty()) {
-            reply = Varicode::rstrip(m_lastTxMessage);
-        }
-
-        // PROCESS ACTIVE HEARTBEAT
-        // if we have hb mode enabled and auto reply enabled <del>and auto ack
-        // enabled and no callsign is selected</del> update: if we're in HB
-        // mode, doesn't matter if a callsign is selected.
-        else if ((d.cmd == " HB" || d.cmd == " HEARTBEAT") &&
-                 canCurrentModeSendHeartbeat() &&
-                 ui->actionModeJS8HB->isChecked() &&
-                 ui->actionModeAutoreply->isChecked() &&
-                 ui->actionHeartbeatAcknowledgements->isChecked()) {
-            // do not process HB activity if buffer is not empty, this prevents
-            // broken incoming MSG's
-            if (!m_messageBuffer.isEmpty()) {
-                qCDebug(mainwindow_js8) << "hb paused for messageBuffer";
-                continue;
-            }
-
-            // check to make sure we aren't pausing HB transmissions (ACKs)
-            // while a callsign is selected
-            if (m_config.heartbeat_qso_pause() && !selectedCallsign.isEmpty()) {
-                qCDebug(mainwindow_js8) << "hb paused during qso";
-                continue;
-            }
-
-            // check to make sure this callsign isn't blacklisted
-            if (m_config.hb_blacklist().contains(d.from) ||
-                m_config.hb_blacklist().contains(
-                    Radio::base_callsign(d.from))) {
-                qCDebug(mainwindow_js8) << "hb blacklist blocking" << d.from;
-                continue;
-            }
-
-            // check to see if we have a message for a station who is
-            // heartbeating
-            QString extra;
-            auto mid = getNextMessageIdForCallsign(d.from);
-            if (mid != -1) {
-                extra = QString("MSG ID %1").arg(mid);
-            }
-
-            // group messaging - if isGroupCall, check to see if there's a
-            // message id for the group and return it if there's not an
-            // individual message
-            // TODO: include group name in response to differentiate from direct
-            // messages?
-            else if (isGroupCall) {
-                mid = getNextGroupMessageIdForCallsign(d.to, d.from);
-                if (mid != -1) {
-                    extra = QString("MSG ID %1").arg(mid);
-                }
-            }
-
-            // TODO: require confirmation?
-            sendHeartbeatAck(d.from, d.snr, extra);
-
-            if (isAllCall) {
-                // since all pings are technically @ALLCALL, let's bump the
-                // allcall cache here...
-                m_txAllcallCommandCache.insert(d.from, new QDateTime(now), 5);
-            }
-
-            continue;
-        }
-
-        // PROCESS HEARTBEAT SNR
-        else if (d.cmd == " HEARTBEAT SNR") {
-            qCDebug(mainwindow_js8) << "skipping incoming hb snr" << d.text;
-            continue;
-        }
-
-        // PROCESS CQ
-        else if (d.cmd == " CQ") {
-            qCDebug(mainwindow_js8) << "skipping incoming cq" << d.text;
-            continue;
-        }
-
-        // PROCESS MSG
-        else if (d.cmd == " MSG" && !isAllCall) {
-
-            auto text = d.text;
-
-            qCDebug(mainwindow_js8) << "adding message to inbox" << text;
-
-            auto calls = parseRelayPathCallsigns(d.from, text);
-
-            d.cmd = " MSG ";
-            d.relayPath = calls.join(">");
-            d.text = text;
-
-            addCommandToMyInbox(d);
-
-            // notification
-            tryNotify("inbox");
-
-            // we haven't replaced the from with the relay path, so we have to
-            // use it for the ack if there is one
-            reply = QString("%1 ACK").arg(calls.length() > 1 ? d.relayPath
-                                                             : d.from);
-
-#define SHOW_ALERT_FOR_MSG 1
-#if SHOW_ALERT_FOR_MSG
-            SelfDestructMessageBox *m = new SelfDestructMessageBox(
-                300, "New Message Received",
-                QString("A new message was received at %1 UTC from %2")
-                    .arg(d.utcTimestamp.time().toString())
-                    .arg(d.from),
-                QMessageBox::Information, QMessageBox::Ok, QMessageBox::Ok,
-                false, this);
-
-            m->show();
-#endif
-        }
-
-        // PROCESS ACKS
-        else if (d.cmd == " ACK" && !isAllCall) {
-            qCDebug(mainwindow_js8) << "skipping incoming ack" << d.text;
-
-            // notification for ack
-            tryNotify("ack");
-
-            // make sure this is explicit
-            continue;
-        }
-
-        // PROCESS BUFFERED CMD
-        else if (d.cmd == " CMD" && !isAllCall) {
-            qCDebug(mainwindow_js8) << "skipping incoming command" << d.text;
-
-            // make sure this is explicit
-            continue;
-        }
-
-        // PROCESS BUFFERED QUERY
-        else if (d.cmd == " QUERY" && !isAllCall) {
-            auto who = d.from; // keep in mind, this is the sender, not the
-                               // original requestor if relayed
-            auto replyPath = d.from;
-
-            if (d.relayPath.contains(">")) {
-                auto path = d.relayPath.split(">");
-                who = path.last();
-                replyPath = d.relayPath;
-            }
-
-            QStringList segs = d.text.split(" ");
-            if (segs.isEmpty()) {
-                continue;
-            }
-
-            auto cmd = segs.first();
-            segs.removeFirst();
-
-            if (cmd == "MSG" && !segs.isEmpty()) {
-                auto inbox = Inbox(inboxPath());
-                if (!inbox.open()) {
-                    continue;
-                }
-
-                bool ok = false;
-                int mid = QString(segs.first()).toInt(&ok);
-                if (!ok) {
-                    continue;
-                }
-
-                auto msg = inbox.value(mid);
-                auto params = msg.params();
-                if (params.isEmpty()) {
-                    continue;
-                }
-
-                auto from = params.value("FROM").toString().trimmed();
-
-                auto to = params.value("TO").toString().trimmed();
-
-                // group messaging - allow any message to a @GROUP to be
-                // retrieved by anybody
-                bool isGroupMsg = to.startsWith("@");
-
-                if (!isGroupMsg && to != who &&
-                    to != Radio::base_callsign(who)) {
-                    continue;
-                }
-
-                auto text = params.value("TEXT").toString().trimmed();
-                if (text.isEmpty()) {
-                    continue;
-                }
-
-                /*
-                 * mark as delivered (so subsequent HBs and QUERY MSGS don't
-                 * receive this message)
-                 *
-                 * Do this in callbacks so that messages are only marked read
-                 * after any potential confirmations have been made by the user,
-                 * and the message has been processed in the transaction queue.
-                 */
-                if (!isGroupMsg) {
-                    callback = [this, mid, msg]() {
-                        this->markMsgDelivered(mid, msg);
-                    };
-                } else {
-                    callback = [this, mid, who]() {
-                        this->markGroupMsgDeliveredForCallsign(mid, who);
-                    };
-                }
-
-                auto lookaheadMid = getLookaheadMessageIdForCallsign(who, mid);
-                if (lookaheadMid == -1 && isGroupMsg) {
-                    lookaheadMid =
-                        getLookaheadGroupMessageIdForCallsign(d.to, who, mid);
-                }
-
-                // and reply
-                if (lookaheadMid != -1) {
-                    reply = QString("%1 MSG %2 FROM %3 NEXT MSG ID %4");
-                    reply = reply.arg(replyPath);
-                    reply = reply.arg(text);
-                    reply = reply.arg(from);
-                    reply = reply.arg(lookaheadMid);
-                } else {
-                    reply = QString("%1 MSG %2 FROM %3");
-                    reply = reply.arg(replyPath);
-                    reply = reply.arg(text);
-                    reply = reply.arg(from);
-                }
-            }
-        }
-
-        // PROCESS BUFFERED QUERY MSGS
-        // Do not process this request if there is incoming MSG as it breaks the
-        // incoming
-        else if (d.cmd == " QUERY MSGS" &&
-                 ui->actionModeAutoreply->isChecked() &&
-                 m_messageBuffer.isEmpty()) {
-            auto who = d.from; // keep in mind, this is the sender, not the
-                               // original requestor if relayed
-            auto replyPath = d.from;
-
-            if (d.relayPath.contains(">")) {
-                auto path = d.relayPath.split(">");
-                who = path.last();
-                replyPath = d.relayPath;
-            }
-
-            // if this is an allcall or a directed call, check to see if we have
-            // a stored message for user. we reply yes if the user would be able
-            // to retreive a stored message
-            auto mid = getNextMessageIdForCallsign(who);
-            if (mid != -1) {
-                reply = QString("%1 YES MSG ID %2").arg(replyPath).arg(mid);
-            }
-
-            // Group messaging - if isGroupCall, check to see if there's a
-            // message id for the group and return it if there's not an
-            // individual message
-            // TODO: include group name in response to differentiate from direct
-            // messages?
-            else if (isGroupCall) {
-                mid = getNextGroupMessageIdForCallsign(d.to, d.from);
-                if (mid != -1) {
-                    reply = QString("%1 YES MSG ID %2").arg(replyPath).arg(mid);
-                }
-            }
-
-            // if this is not an allcall and we have no messages, reply no.
-            if (!isAllCall && reply.isEmpty()) {
-                reply = QString("%1 NO").arg(replyPath);
-            }
-        }
-
-        // PROCESS BUFFERED QUERY CALL
-        // Do not process this request if there is incoming MSG as it breaks the
-        // incoming
-        else if (d.cmd == " QUERY CALL" &&
-                 ui->actionModeAutoreply->isChecked() &&
-                 m_messageBuffer.isEmpty()) {
-            auto replyPath = d.from;
-            if (d.relayPath.contains(">")) {
-                replyPath = d.relayPath;
-            }
-
-            auto who = d.text;
-            if (who.isEmpty()) {
-                continue;
-            }
-
-            auto callsigns = Varicode::parseCallsigns(who);
-            if (callsigns.isEmpty()) {
-                continue;
-            }
-
-            QStringList replies;
-            int callsignAging = m_config.callsign_aging();
-            auto baseCall = callsigns.first();
-            foreach (auto cd, m_callActivity.values()) {
-                if (callsignAging &&
-                    cd.utcTimestamp.secsTo(now) / 60 >= callsignAging) {
-                    continue;
-                }
-
-                if (baseCall == cd.call ||
-                    baseCall == Radio::base_callsign(cd.call)) {
-                    auto r = QString("%1 (%2)")
-                                 .arg(Varicode::formatSNR(cd.snr))
-                                 .arg(since(cd.utcTimestamp))
-                                 .trimmed();
-                    replies.append(r);
-                    break;
-                }
-            }
-
-            if (!replies.isEmpty()) {
-                replies.prepend(QString("%1 YES").arg(replyPath));
-            }
-
-            reply = replies.join(" ");
-
-            if (!reply.isEmpty()) {
-                if (isAllCall) {
-                    m_txAllcallCommandCache.insert(d.from, new QDateTime(now),
-                                                   25);
-                }
-            }
-        }
-
-#if 0
-        // PROCESS ALERT
-        else if (d.cmd == "!" && !isAllCall) {
-
-            // create alert dialog
-            processAlertReplyForCommand(d, d.from, " ");
-
-            // make sure this is explicit
-            continue;
-        }
-#endif
-
-        // well, if there's no reply, don't do anything...
-        if (reply.isEmpty()) {
-            continue;
-        }
-
-        // do not queue @ALLCALL replies if auto-reply is not checked
-        if (!ui->actionModeAutoreply->isChecked() && isAllCall) {
-            continue;
-        }
-
-#if 0
-        // TODO: jsherer - HB issue here
-        // do not queue a reply if it's a HB and HB is not active
-        // if((!ui->hbMacroButton->isChecked() || m_hbInterval <= 0) && d.cmd.contains("HB")){
-        //     continue;
-        // }
-#endif
-
-        // do not queue for reply if there's text in the window
-        if (!ui->extFreeTextMsgEdit->toPlainText().isEmpty()) {
-            continue;
-        }
-
-        // do not queue for reply if there's a buffer open to us
-        int bufferOffset = 0;
-        if (hasExistingMessageBufferToMe(&bufferOffset)) {
-            qCDebug(mainwindow_js8) << "skipping reply due to open buffer"
-                                    << bufferOffset << m_messageBuffer.count();
-            continue;
-        }
-
-        // add @ALLCALLs to the @ALLCALL cache
-        if (isAllCall) {
-            m_txAllcallCommandCache.insert(d.from, new QDateTime(now), 25);
-        }
-
-        // queue the reply here to be sent when a free interval is available on
-        // the frequency that was sent unless, this is an allcall, to which we
-        // should be responding on a clear frequency offset we always want to
-        // make sure that the directed cache has been updated at this point so
-        // we have the most information available to make a frequency selection.
-        if (m_config.autoreply_confirmation()) {
-            confirmThenEnqueueMessage(90, priority, reply, freq, callback);
-        } else {
-            enqueueMessage(priority, reply, freq, callback);
-        }
-    }
-}
+void processCommandActivity(); // JS8_Mainwindow/processCommandActivity.cpp
 
 QString MainWindow::inboxPath() {
     return QDir::toNativeSeparators(
@@ -10779,8 +7831,8 @@ void MainWindow::processTxQueue() {
         return;
     }
 
-    // and if we are a low priority message, we need to have not transmitted in
-    // the past 30 seconds...
+    // and if we are a low priority message, we need to have not transmitted
+    // in the past 30 seconds...
     if (head.priority <= PriorityLow &&
         m_lastTxStartTime.secsTo(DriftingDateTime::currentDateTimeUtc()) <=
             30) {
@@ -10793,8 +7845,9 @@ void MainWindow::processTxQueue() {
     // add the message to the outgoing message text box
     addMessageText(message.message, true);
 
-    // check to see if this is a high priority message, or if we have autoreply
-    // enabled, or if this is a ping and the ping button is enabled
+    // check to see if this is a high priority message, or if we have
+    // autoreply enabled, or if this is a ping and the ping button is
+    // enabled
     if (message.priority >= PriorityHigh ||
         message.message.contains(" HEARTBEAT ") ||
         message.message.contains(" HB ") || message.message.contains(" ACK ") ||
@@ -10826,897 +7879,10 @@ void MainWindow::displayActivity(bool force) {
 }
 
 // updateBandActivity
-void MainWindow::displayBandActivity() {
-    auto now = DriftingDateTime::currentDateTimeUtc();
-
-    // Reset the header label text to accommodate minimal label setting
-    int cols = ui->tableWidgetRXAll->columnCount();
-    for (int c = 0; c < cols; ++c) {
-        ui->tableWidgetRXAll->horizontalHeaderItem(c)->setText(
-            columnLabel(m_origRxHeaderLabelMap[c]));
-    }
-
-    ui->tableWidgetRXAll->setFont(m_config.table_font());
-
-    // Selected Offset
-    int selectedOffset = -1;
-    auto selectedItems = ui->tableWidgetRXAll->selectedItems();
-    if (!selectedItems.isEmpty()) {
-        selectedOffset = selectedItems.first()->data(Qt::UserRole).toInt();
-    }
-
-    ui->tableWidgetRXAll->setUpdatesEnabled(false);
-    {
-        // Scroll Position
-        auto const currentScrollPos =
-            ui->tableWidgetRXAll->verticalScrollBar()->value();
-
-        // Clear the table
-        ui->tableWidgetRXAll->setRowCount(0);
-
-        // Sort!
-        auto const sort = getSortByReverse("bandActivity", "offset");
-        auto keys = m_bandActivity.keys();
-
-        // Base comparison, called by the detail comparisons. We may not need
-        // to proceed to the detail comparison at all here, if at least one of
-        // the lists is empty. If both are non-empty, delegate to the detail
-        // comparison, providing it with the last list elements.
-
-        auto const compare = [this](int const lhsKey, int const rhsKey,
-                                    auto &&detail) {
-            auto const &lhs = m_bandActivity[lhsKey];
-            auto const &rhs = m_bandActivity[rhsKey];
-
-            if (lhs.isEmpty())
-                return false;
-            if (rhs.isEmpty())
-                return true;
-
-            return detail(lhs.last(), rhs.last());
-        };
-
-        // Time stamp comparison, easy stuff, just a total ordering on the
-        // UTC time stamp field.
-
-        auto const compareTimestamp = [compare](int const lhsKey,
-                                                int const rhsKey) {
-            return compare(lhsKey, rhsKey, [](auto &&lhs, auto &&rhs) {
-                return lhs.utcTimestamp < rhs.utcTimestamp;
-            });
-        };
-
-        // SNR comparison;  we always want insane SNR values to be at the end
-        // of the list and the list is going to be reversed if reverse is set,
-        // so we want to set things up so that insane elements are either all
-        // at the beginning in the case of a reverse, or all at the end in the
-        // standard case. Reverse takes care of itself; we just need to sort
-        // out standard.
-
-        auto const compareSNR = [compare, reverse = sort.reverse](
-                                    int const lhsKey, int const rhsKey) {
-            return compare(lhsKey, rhsKey, [reverse](auto &&lhs, auto &&rhs) {
-                auto lhsSNR = lhs.snr;
-                auto rhsSNR = rhs.snr;
-
-                if (!reverse) {
-                    if (lhsSNR < -60 || lhsSNR > 60)
-                        lhsSNR = -lhsSNR;
-                    if (rhsSNR < -60 || rhsSNR > 60)
-                        rhsSNR = -rhsSNR;
-                }
-
-                return lhsSNR < rhsSNR;
-            });
-        };
-
-        // Submode comparison; slow mode isn't at the start of the enumeration;
-        // it's in the middle of it. All the other modes are in the expected
-        // order.
-
-        auto const compareSubmode = [compare](int const lhsKey,
-                                              int const rhsKey) {
-            return compare(lhsKey, rhsKey, [](auto &&lhs, auto &&rhs) {
-                auto lhsSubmode = lhs.submode;
-                auto rhsSubmode = rhs.submode;
-
-                if (lhsSubmode == Varicode::JS8CallSlow)
-                    lhsSubmode = -lhsSubmode;
-                if (rhsSubmode == Varicode::JS8CallSlow)
-                    rhsSubmode = -rhsSubmode;
-
-                return lhsSubmode < rhsSubmode;
-            });
-        };
-
-        // Always perform an initial sort by offset.
-
-        std::stable_sort(keys.begin(), keys.end());
-
-        // If something other than offset was requested as the sort by, perform
-        // an additional stable sort by the field requested.
-
-        if (sort.by == "timestamp")
-            std::stable_sort(keys.begin(), keys.end(), compareTimestamp);
-        else if (sort.by == "snr")
-            std::stable_sort(keys.begin(), keys.end(), compareSNR);
-        else if (sort.by == "submode")
-            std::stable_sort(keys.begin(), keys.end(), compareSubmode);
-
-        // The sort comparators leave things in forward order. If a reverse sort
-        // was requested, reverse the keys.
-
-        if (sort.reverse)
-            std::reverse(keys.begin(), keys.end());
-
-        // Build the table
-        foreach (int offset, keys) {
-            bool isOffsetSelected = (offset == selectedOffset);
-
-            QList<ActivityDetail> items = m_bandActivity[offset];
-            if (items.length() > 0) {
-                QDateTime timestamp;
-                QStringList text;
-                QString age;
-                int snr = 0;
-                float tdrift = 0;
-                int submode = -1;
-
-                int activityAging = m_config.activity_aging();
-
-                // hide items that shouldn't appear
-
-                for (int i = 0; i < items.length(); i++) {
-                    auto item = items[i];
-
-                    bool shouldDisplay = true;
-
-                    // hide aged items
-                    if (!isOffsetSelected && activityAging &&
-                        item.utcTimestamp.secsTo(now) / 60 >= activityAging) {
-                        shouldDisplay = false;
-                    }
-
-                    // hide heartbeat items
-                    if (!ui->actionShow_Band_Heartbeats_and_ACKs->isChecked()) {
-                        // hide heartbeats and acks if we have heartbeating
-                        // hidden
-                        if (item.text.contains(" @HB ") ||
-                            item.text.contains(" HEARTBEAT ")) {
-                            shouldDisplay = false;
-
-                            // hide the previous item if this it shouldn't be
-                            // displayed either...
-                            if (i > 0 && items[i - 1].shouldDisplay &&
-                                items[i - 1].text.endsWith(": ")) {
-                                items[i - 1].shouldDisplay = false;
-                            }
-                        }
-
-                        // if our previous item should not be displayed (or this
-                        // is the first frame) and we have a MSG ID, then don't
-                        // display it either.
-                        if ((i == 0 ||
-                             (i > 0 && !items[i - 1].shouldDisplay)) &&
-                            (item.text.contains(" MSG ID "))) {
-                            shouldDisplay = false;
-                        }
-                    }
-
-                    // hide empty items
-                    if (item.text.isEmpty()) {
-                        shouldDisplay = false;
-                    }
-
-                    // set the visibility of the item
-                    items[i].shouldDisplay = shouldDisplay;
-                }
-
-                // show the items that should appear
-                foreach (ActivityDetail item, items) {
-                    if (!item.shouldDisplay) {
-                        continue;
-                    }
-
-                    if (item.isLowConfidence) {
-                        item.text = QString("[%1]").arg(item.text);
-                    }
-
-                    if ((item.bits & Varicode::JS8CallLast) ==
-                        Varicode::JS8CallLast) {
-                        // append the eot character to the text
-                        item.text = QString("%1 %2 ")
-                                        .arg(Varicode::rstrip(item.text))
-                                        .arg(m_config.eot());
-                    }
-                    text.append(item.text);
-                    snr = item.snr;
-                    age = since(item.utcTimestamp);
-                    timestamp = item.utcTimestamp;
-                    tdrift = item.tdrift;
-                    submode = item.submode;
-                }
-
-                auto joined = Varicode::rstrip(text.join(""));
-                if (joined.isEmpty()) {
-                    continue;
-                }
-
-                ui->tableWidgetRXAll->insertRow(
-                    ui->tableWidgetRXAll->rowCount());
-                int row = ui->tableWidgetRXAll->rowCount() - 1;
-                int col = 0;
-
-                auto offsetItem = new QTableWidgetItem(
-                    QString(columnLabel("%1 Hz")).arg(offset));
-                offsetItem->setData(Qt::UserRole, QVariant(offset));
-                offsetItem->setTextAlignment(Qt::AlignRight | Qt::AlignVCenter);
-                ui->tableWidgetRXAll->setItem(row, col++, offsetItem);
-
-                auto ageItem = new QTableWidgetItem(age);
-                ageItem->setTextAlignment(Qt::AlignCenter);
-                ageItem->setToolTip(timestamp.toString());
-                ui->tableWidgetRXAll->setItem(row, col++, ageItem);
-
-                auto snrText = Varicode::formatSNR(snr);
-                auto snrItem = new QTableWidgetItem(
-                    snrText.isEmpty()
-                        ? ""
-                        : QString(columnLabel("%1 dB")).arg(snrText));
-                snrItem->setTextAlignment(Qt::AlignRight | Qt::AlignVCenter);
-                ui->tableWidgetRXAll->setItem(row, col++, snrItem);
-
-                auto tdriftItem = new QTableWidgetItem(
-                    QString(columnLabel("%1 ms")).arg((int)(1000 * tdrift)));
-                tdriftItem->setData(Qt::UserRole, QVariant(tdrift));
-                tdriftItem->setTextAlignment(Qt::AlignRight | Qt::AlignVCenter);
-                ui->tableWidgetRXAll->setItem(row, col++, tdriftItem);
-
-                auto name = JS8::Submode::name(submode);
-                auto submodeItem =
-                    new QTableWidgetItem(name.left(1).replace("H", "N"));
-                submodeItem->setToolTip(name);
-                submodeItem->setData(Qt::UserRole, QVariant(name));
-                submodeItem->setTextAlignment(Qt::AlignCenter);
-                ui->tableWidgetRXAll->setItem(row, col++, submodeItem);
-
-                // align right if eliding...
-                int colWidth = ui->tableWidgetRXAll->columnWidth(3);
-                auto textItem = new QTableWidgetItem(joined);
-                auto html = QString("<qt/>%1").arg(joined.toHtmlEscaped());
-                html =
-                    html.replace(m_config.eot(), m_config.eot() + "<br/><br/>");
-                html = html.replace(QRegularExpression("([<]br[/][>])+$"), "");
-                textItem->setToolTip(html);
-
-                QFontMetrics fm(textItem->font());
-                auto elidedText =
-                    fm.elidedText(joined, Qt::ElideLeft, colWidth);
-                auto flag = Qt::AlignLeft | Qt::AlignVCenter;
-                if (elidedText != joined) {
-                    flag = Qt::AlignRight | Qt::AlignVCenter;
-                    textItem->setText(joined);
-                }
-                textItem->setTextAlignment(flag);
-
-                ui->tableWidgetRXAll->setItem(row, col++, textItem);
-
-                if (isOffsetSelected) {
-                    for (int i = 0; i < ui->tableWidgetRXAll->columnCount();
-                         i++) {
-                        ui->tableWidgetRXAll->item(row, i)->setSelected(true);
-                    }
-                }
-
-                bool isDirectedAllCall = false;
-                if ((isDirectedOffset(offset, &isDirectedAllCall) &&
-                     !isDirectedAllCall) ||
-                    isMyCallIncluded(text.last())) {
-                    for (int i = 0; i < ui->tableWidgetRXAll->columnCount();
-                         i++) {
-                        ui->tableWidgetRXAll->item(row, i)->setBackground(
-                            QBrush(m_config.color_MyCall()));
-                    }
-                }
-
-                if (!text.isEmpty()) {
-                    auto const list = joined.split(QRegularExpression("[:> ]"),
-                                                   Qt::SkipEmptyParts);
-                    QSet<QString> words(list.begin(), list.end());
-
-                    if (words.contains("CQ")) {
-                        for (int i = 0; i < ui->tableWidgetRXAll->columnCount();
-                             i++) {
-                            ui->tableWidgetRXAll->item(row, i)->setBackground(
-                                QBrush(m_config.color_CQ()));
-                        }
-                    }
-
-                    auto matchingSecondaryWords =
-                        m_config.secondary_highlight_words() & words;
-                    if (!matchingSecondaryWords.isEmpty()) {
-                        for (int i = 0; i < ui->tableWidgetRXAll->columnCount();
-                             i++) {
-                            ui->tableWidgetRXAll->item(row, i)->setBackground(
-                                QBrush(m_config.color_secondary_highlight()));
-                        }
-                    }
-
-                    auto matchingPrimaryWords =
-                        m_config.primary_highlight_words() & words;
-                    if (!matchingPrimaryWords.isEmpty()) {
-                        for (int i = 0; i < ui->tableWidgetRXAll->columnCount();
-                             i++) {
-                            ui->tableWidgetRXAll->item(row, i)->setBackground(
-                                QBrush(m_config.color_primary_highlight()));
-                        }
-                    }
-                }
-            }
-        }
-
-        // Set table color
-        auto style = QString(
-            "QTableWidget { background:%1; selection-background-color:%2; "
-            "alternate-background-color:%1; color:%3; } "
-            "QTableWidget::item:selected { background-color: %2; color: %3; }");
-        style = style.arg(m_config.color_table_background().name());
-        style = style.arg(m_config.color_table_highlight().name());
-        style = style.arg(m_config.color_table_foreground().name());
-        ui->tableWidgetRXAll->setStyleSheet(style);
-
-        // Set the table palette for inactive selected row
-        auto p = ui->tableWidgetRXAll->palette();
-
-        p.setColor(QPalette::Highlight, m_config.color_table_highlight());
-        p.setColor(QPalette::HighlightedText,
-                   m_config.color_table_foreground());
-        p.setColor(QPalette::Inactive, QPalette::Highlight,
-                   p.color(QPalette::Active, QPalette::Highlight));
-        ui->tableWidgetRXAll->setPalette(p);
-
-        // Set item fonts
-        for (int row = 0; row < ui->tableWidgetRXAll->rowCount(); row++) {
-            for (int col = 0; col < ui->tableWidgetRXAll->columnCount();
-                 col++) {
-                auto item = ui->tableWidgetRXAll->item(row, col);
-                if (item) {
-                    item->setFont(m_config.table_font());
-                }
-            }
-        }
-
-        // Column labels
-        ui->tableWidgetRXAll->horizontalHeader()->setVisible(
-            showColumn("band", "labels"));
-
-        // Hide columns
-        ui->tableWidgetRXAll->setColumnHidden(0, !showColumn("band", "offset"));
-        ui->tableWidgetRXAll->setColumnHidden(1,
-                                              !showColumn("band", "timestamp"));
-        ui->tableWidgetRXAll->setColumnHidden(2, !showColumn("band", "snr"));
-        ui->tableWidgetRXAll->setColumnHidden(
-            3, !showColumn("band", "tdrift", false));
-        ui->tableWidgetRXAll->setColumnHidden(
-            4, !showColumn("band", "submode", false));
-
-        // Resize the table columns
-        ui->tableWidgetRXAll->resizeColumnToContents(0);
-        ui->tableWidgetRXAll->resizeColumnToContents(1);
-        ui->tableWidgetRXAll->resizeColumnToContents(2);
-        ui->tableWidgetRXAll->resizeColumnToContents(3);
-        ui->tableWidgetRXAll->resizeColumnToContents(4);
-
-        // Reset the scroll position
-        ui->tableWidgetRXAll->verticalScrollBar()->setValue(currentScrollPos);
-    }
-    ui->tableWidgetRXAll->setUpdatesEnabled(true);
-}
+void displayBandActivity(); // JS8_Mainwindow/displayBandActivity.cpp
 
 // updateCallActivity
-void MainWindow::displayCallActivity() {
-    auto now = DriftingDateTime::currentDateTimeUtc();
-
-    // Reset the header label text to accommodate minimal label setting
-    int cols = ui->tableWidgetCalls->columnCount();
-    for (int c = 0; c < cols; ++c) {
-        ui->tableWidgetCalls->horizontalHeaderItem(c)->setText(
-            columnLabel(m_origCallActivityHeaderLabelMap[c]));
-    }
-
-    ui->tableWidgetCalls->setFont(m_config.table_font());
-
-    // Selected callsign
-    QString selectedCall = callsignSelected();
-
-    auto currentScrollPos = ui->tableWidgetCalls->verticalScrollBar()->value();
-
-    ui->tableWidgetCalls->setUpdatesEnabled(false);
-    {
-        // Clear the table
-        ui->tableWidgetCalls->setRowCount(0);
-        ui->tableWidgetCalls->horizontalHeaderItem(8)->setText(
-            m_config.miles() ? "mi" : "km");
-
-        bool showIconColumn = false;
-        createGroupCallsignTableRows(
-            ui->tableWidgetCalls, selectedCall,
-            showIconColumn); // isAllCallIncluded(selectedCall)); // ||
-                             // isGroupCallIncluded(selectedCall));
-
-        // Build the table
-
-        auto const sort = getSortByReverse("callActivity", "callsign");
-        auto keys = m_callActivity.keys();
-
-        auto const compareOffset = [this](QString const &lhsKey,
-                                          QString const &rhsKey) {
-            return m_callActivity[lhsKey].offset <
-                   m_callActivity[rhsKey].offset;
-        };
-
-        auto const compareAzimuth =
-            [this, reverse = sort.reverse, my_grid = m_config.my_grid()](
-                QString const &lhsKey, QString const &rhsKey) {
-                auto const lhs =
-                    Geodesic::vector(my_grid, m_callActivity[lhsKey].grid)
-                        .azimuth();
-                auto const rhs =
-                    Geodesic::vector(my_grid, m_callActivity[rhsKey].grid)
-                        .azimuth();
-
-                // We always want invalid azimuths to be at the end of the list,
-                // and the list is going to be reversed if reverse is set, so we
-                // want to set things up so that invalid elements are either all
-                // at the beginning in the case of a reverse, or all at the end
-                // in the standard case.
-
-                if (!lhs)
-                    return reverse && rhs;
-                else if (!rhs)
-                    return !reverse;
-                else
-                    return lhs < rhs;
-            };
-
-        auto const compareDistance =
-            [this, reverse = sort.reverse, my_grid = m_config.my_grid()](
-                QString const &lhsKey, QString const &rhsKey) {
-                auto const lhs =
-                    Geodesic::vector(my_grid, m_callActivity[lhsKey].grid)
-                        .distance();
-                auto const rhs =
-                    Geodesic::vector(my_grid, m_callActivity[rhsKey].grid)
-                        .distance();
-
-                // We always want invalid distances to be at the end of the
-                // list, and the list is going to be reversed if reverse is set,
-                // so we want to set things up so that invalid elements are
-                // either all at the beginning in the case of a reverse, or all
-                // at the end in the standard case.
-
-                if (!lhs)
-                    return reverse && rhs;
-                else if (!rhs)
-                    return !reverse;
-                else
-                    return lhs < rhs;
-            };
-
-        auto const compareTimestamp = [this](QString const &lhsKey,
-                                             QString const &rhsKey) {
-            return m_callActivity[lhsKey].utcTimestamp <
-                   m_callActivity[rhsKey].utcTimestamp;
-        };
-
-        auto const compareAckTimestamp = [this](QString const &lhsKey,
-                                                QString const &rhsKey) {
-            return m_callActivity[rhsKey].ackTimestamp <
-                   m_callActivity[lhsKey].ackTimestamp;
-        };
-
-        auto const compareSNR =
-            [this, reverse = sort.reverse](QString const &lhsKey,
-                                           QString const &rhsKey) {
-                auto lhs = m_callActivity[lhsKey].snr;
-                auto rhs = m_callActivity[rhsKey].snr;
-
-                // We always want insane SNR values to be at the end of the
-                // list, and the list is going to be reversed if reverse is set,
-                // so we want to set things up so that insane elements are
-                // either all at the beginning in the case of a reverse, or all
-                // at the end in the standard case. Reverse takes care of
-                // itself; we just need to sort out standard.
-
-                if (!reverse) {
-                    if (lhs < -60 || lhs > 60)
-                        lhs = -lhs;
-                    if (rhs < -60 || rhs > 60)
-                        rhs = -rhs;
-                }
-
-                return lhs < rhs;
-            };
-
-        auto const compareSubmode = [this](QString const &lhsKey,
-                                           QString const &rhsKey) {
-            auto lhs = m_callActivity[lhsKey].submode;
-            auto rhs = m_callActivity[rhsKey].submode;
-
-            // Slow mode isn't at the start of the enumeration; it's in the
-            // middle of it. All the other modes are in the expected order.
-
-            if (lhs == Varicode::JS8CallSlow)
-                lhs = -lhs;
-            if (rhs == Varicode::JS8CallSlow)
-                rhs = -rhs;
-
-            return lhs < rhs;
-        };
-
-        // Always perform an initial sort by callsign.
-
-        std::stable_sort(keys.begin(), keys.end());
-
-        // If something other than callsign was requested as the sort by,
-        // perform an additional stable sort by the field requested.
-
-        if (sort.by == "offset")
-            std::stable_sort(keys.begin(), keys.end(), compareOffset);
-        else if (sort.by == "distance")
-            std::stable_sort(keys.begin(), keys.end(), compareDistance);
-        else if (sort.by == "azimuth")
-            std::stable_sort(keys.begin(), keys.end(), compareAzimuth);
-        else if (sort.by == "timestamp")
-            std::stable_sort(keys.begin(), keys.end(), compareTimestamp);
-        else if (sort.by == "ackTimestamp")
-            std::stable_sort(keys.begin(), keys.end(), compareAckTimestamp);
-        else if (sort.by == "snr")
-            std::stable_sort(keys.begin(), keys.end(), compareSNR);
-        else if (sort.by == "submode")
-            std::stable_sort(keys.begin(), keys.end(), compareSubmode);
-
-        // The sort comparators leave things in forward order. If a reverse sort
-        // was requested, reverse the keys.
-
-        if (sort.reverse)
-            std::reverse(keys.begin(), keys.end());
-
-        // pin messages to the top
-        std::stable_sort(keys.begin(), keys.end(),
-                         [this](QString const &lhsKey, QString const &rhsKey) {
-                             auto const lhs = (int)!(
-                                 m_rxInboxCountCache.value(lhsKey, 0) > 0);
-                             auto const rhs = (int)!(
-                                 m_rxInboxCountCache.value(rhsKey, 0) > 0);
-
-                             return lhs < rhs;
-                         });
-
-        int callsignAging = m_config.callsign_aging();
-        foreach (QString call, keys) {
-            if (call.trimmed().isEmpty()) {
-                continue;
-            }
-
-            CallDetail d = m_callActivity[call];
-            if (d.call.trimmed().isEmpty()) {
-                continue;
-            }
-
-            bool isCallSelected = (call == selectedCall);
-
-            // icon flags (flag -> star -> empty)
-            bool hasMessage = m_rxInboxCountCache.value(d.call, 0) > 0;
-
-            // display telephone icon if called cq in the past 5 minutes
-            bool hasCQ =
-                d.cqTimestamp.isValid() && d.cqTimestamp.secsTo(now) / 60 < 5;
-
-            // display star if they've acked a message from us
-            bool hasACK = d.ackTimestamp.isValid();
-
-            if (!isCallSelected && !hasMessage && callsignAging &&
-                d.utcTimestamp.secsTo(now) / 60 >= callsignAging) {
-                continue;
-            }
-
-            ui->tableWidgetCalls->insertRow(ui->tableWidgetCalls->rowCount());
-            int row = ui->tableWidgetCalls->rowCount() - 1;
-            int col = 0;
-
-#if SHOW_THROUGH_CALLS
-            QString displayCall =
-                d.through.isEmpty()
-                    ? d.call
-                    : QString("%1>%2").arg(d.through).arg(d.call);
-#else
-            QString displayCall = d.call;
-#endif
-            bool hasThrough = !d.through.isEmpty();
-
-            auto iconItem = new QTableWidgetItem(hasMessage   ? "\u2691"
-                                                 : hasACK     ? "\u2605"
-                                                 : hasCQ      ? "\u260E"
-                                                 : hasThrough ? "\u269F"
-                                                              : "");
-            iconItem->setData(Qt::UserRole, QVariant(d.call));
-            iconItem->setToolTip(
-                hasMessage ? "Message Available"
-                : hasACK   ? QString("Hearing Your Station (%1)")
-                               .arg(since(d.ackTimestamp))
-                : hasCQ ? QString("Calling CQ (%1)").arg(since(d.cqTimestamp))
-                : hasThrough
-                    ? QString("Heard Through Relay (%1)").arg(d.through)
-                    : "");
-            iconItem->setTextAlignment(Qt::AlignCenter);
-            ui->tableWidgetCalls->setItem(row, col++, iconItem);
-            if (hasMessage || hasACK || hasCQ || hasThrough) {
-                showIconColumn = true;
-            }
-
-            auto displayItem = new QTableWidgetItem(displayCall);
-            displayItem->setData(Qt::UserRole, QVariant(d.call));
-            displayItem->setToolTip(generateCallDetail(displayCall));
-            ui->tableWidgetCalls->setItem(row, col++, displayItem);
-
-#if ONLY_SHOW_HEARD_CALLSIGNS
-            if (d.utcTimestamp.isValid()) {
-#else
-            if (true) {
-#endif
-                auto ageItem = new QTableWidgetItem(since(d.utcTimestamp));
-                ageItem->setTextAlignment(Qt::AlignCenter);
-                ageItem->setToolTip(d.utcTimestamp.toString());
-                ui->tableWidgetCalls->setItem(row, col++, ageItem);
-
-                auto snrText = Varicode::formatSNR(d.snr);
-                auto snrItem = new QTableWidgetItem(
-                    snrText.isEmpty()
-                        ? ""
-                        : QString(columnLabel("%1 dB")).arg(snrText));
-                snrItem->setTextAlignment(Qt::AlignRight | Qt::AlignVCenter);
-                ui->tableWidgetCalls->setItem(row, col++, snrItem);
-
-                auto offsetItem = new QTableWidgetItem(
-                    QString(columnLabel("%1 Hz")).arg(d.offset));
-                offsetItem->setData(Qt::UserRole, QVariant(d.offset));
-                offsetItem->setTextAlignment(Qt::AlignRight | Qt::AlignVCenter);
-                ui->tableWidgetCalls->setItem(row, col++, offsetItem);
-
-                auto tdriftItem = new QTableWidgetItem(
-                    QString(columnLabel("%1 ms")).arg((int)(1000 * d.tdrift)));
-                tdriftItem->setTextAlignment(Qt::AlignRight | Qt::AlignVCenter);
-                ui->tableWidgetCalls->setItem(row, col++, tdriftItem);
-
-                auto name = JS8::Submode::name(d.submode);
-                auto modeItem =
-                    new QTableWidgetItem(name.left(1).replace("H", "N"));
-                modeItem->setToolTip(name);
-                modeItem->setData(Qt::UserRole, QVariant(name));
-                modeItem->setTextAlignment(Qt::AlignCenter);
-                ui->tableWidgetCalls->setItem(row, col++, modeItem);
-
-                auto gridItem = new QTableWidgetItem(
-                    QString("%1").arg(d.grid.trimmed().left(4)));
-                gridItem->setToolTip(d.grid.trimmed());
-                ui->tableWidgetCalls->setItem(row, col++, gridItem);
-
-                auto const vector =
-                    Geodesic::vector(m_config.my_grid(), d.grid);
-                auto const units = !showColumn("call", "labels");
-
-                auto distanceItem = new QTableWidgetItem(
-                    vector.distance().toString(m_config.miles(), units));
-                distanceItem->setTextAlignment(Qt::AlignRight |
-                                               Qt::AlignVCenter);
-                ui->tableWidgetCalls->setItem(row, col++, distanceItem);
-
-                auto azimuthItem =
-                    new QTableWidgetItem(vector.azimuth().toString(units));
-                if (auto const azimuth = vector.azimuth())
-                    azimuthItem->setToolTip(azimuth.compass().toString());
-                azimuthItem->setTextAlignment(Qt::AlignRight |
-                                              Qt::AlignVCenter);
-                ui->tableWidgetCalls->setItem(row, col++, azimuthItem);
-
-                QString flag;
-                if (m_logBook.hasWorkedBefore(d.call, "")) {
-                    // unicode checkmark
-                    flag = "\u2713";
-                }
-                auto workedBeforeItem = new QTableWidgetItem(flag);
-                workedBeforeItem->setTextAlignment(Qt::AlignCenter);
-                ui->tableWidgetCalls->setItem(row, col++, workedBeforeItem);
-
-                QString logDetailGrid;
-                QString logDetailDate;
-                QString logDetailName;
-                QString logDetailComment;
-                bool gridItemEmpty = gridItem->text().isEmpty();
-
-                if ((gridItemEmpty && showColumn("call", "grid")) ||
-                    showColumn("call", "log") ||
-                    showColumn("call", "logName") ||
-                    showColumn("call", "logComment")) {
-                    m_logBook.findCallDetails(d.call, logDetailGrid,
-                                              logDetailDate, logDetailName,
-                                              logDetailComment);
-                }
-
-                if (gridItemEmpty && !logDetailGrid.isEmpty()) {
-                    gridItem->setText(logDetailGrid.trimmed().left(4));
-                    gridItem->setToolTip(logDetailGrid.trimmed());
-
-                    auto const vector =
-                        Geodesic::vector(m_config.my_grid(), d.grid);
-                    auto const units = !showColumn("call", "labels");
-
-                    distanceItem->setText(
-                        vector.distance().toString(m_config.miles(), units));
-                    azimuthItem->setText(vector.azimuth().toString(units));
-                    if (auto const azimuth = vector.azimuth())
-                        azimuthItem->setToolTip(azimuth.compass().toString());
-
-                    // update the call activity cache with the loaded grid
-                    if (m_callActivity.contains(d.call)) {
-                        m_callActivity[call].grid = logDetailGrid.trimmed();
-                    }
-                }
-
-                if (!logDetailDate.isEmpty()) {
-                    auto lastLogged =
-                        QDate::fromString(logDetailDate, "yyyyMMdd");
-
-                    workedBeforeItem->setToolTip(
-                        QString("Last Logged: %1").arg(lastLogged.toString()));
-                }
-
-                auto logNameItem = new QTableWidgetItem(logDetailName);
-                logNameItem->setTextAlignment(Qt::AlignCenter);
-                logNameItem->setToolTip(logDetailName);
-                ui->tableWidgetCalls->setItem(row, col++, logNameItem);
-
-                auto logCommentItem = new QTableWidgetItem(logDetailComment);
-                logCommentItem->setTextAlignment(Qt::AlignCenter);
-                logCommentItem->setToolTip(logDetailComment);
-                ui->tableWidgetCalls->setItem(row, col++, logCommentItem);
-
-            } else {
-                ui->tableWidgetCalls->setItem(row, col++,
-                                              new QTableWidgetItem("")); // age
-                ui->tableWidgetCalls->setItem(row, col++,
-                                              new QTableWidgetItem("")); // snr
-                ui->tableWidgetCalls->setItem(row, col++,
-                                              new QTableWidgetItem("")); // freq
-                ui->tableWidgetCalls->setItem(
-                    row, col++, new QTableWidgetItem("")); // tdrift
-                ui->tableWidgetCalls->setItem(row, col++,
-                                              new QTableWidgetItem("")); // mode
-                ui->tableWidgetCalls->setItem(row, col++,
-                                              new QTableWidgetItem("")); // grid
-                ui->tableWidgetCalls->setItem(
-                    row, col++, new QTableWidgetItem("")); // distance
-                ui->tableWidgetCalls->setItem(
-                    row, col++, new QTableWidgetItem("")); // azimuth
-                ui->tableWidgetCalls->setItem(
-                    row, col++, new QTableWidgetItem("")); // worked before
-                ui->tableWidgetCalls->setItem(
-                    row, col++, new QTableWidgetItem("")); // log name
-                ui->tableWidgetCalls->setItem(
-                    row, col++, new QTableWidgetItem("")); // log comment
-            }
-
-            if (isCallSelected) {
-                for (int i = 0; i < ui->tableWidgetCalls->columnCount(); i++) {
-                    ui->tableWidgetCalls->item(row, i)->setSelected(true);
-                }
-            }
-
-            if (hasCQ) {
-                for (int i = 0; i < ui->tableWidgetCalls->columnCount(); i++) {
-                    ui->tableWidgetCalls->item(row, i)->setBackground(
-                        QBrush(m_config.color_CQ()));
-                }
-            }
-
-            if (m_config.secondary_highlight_words().contains(call)) {
-                for (int i = 0; i < ui->tableWidgetCalls->columnCount(); i++) {
-                    ui->tableWidgetCalls->item(row, i)->setBackground(
-                        QBrush(m_config.color_secondary_highlight()));
-                }
-            }
-
-            if (m_config.primary_highlight_words().contains(call)) {
-                for (int i = 0; i < ui->tableWidgetCalls->columnCount(); i++) {
-                    ui->tableWidgetCalls->item(row, i)->setBackground(
-                        QBrush(m_config.color_primary_highlight()));
-                }
-            }
-        }
-
-        // Set table color
-        auto style = QString(
-            "QTableWidget { background:%1; selection-background-color:%2; "
-            "alternate-background-color:%1; color:%3; } "
-            "QTableWidget::item:selected { background-color: %2; color: %3; }");
-        style = style.arg(m_config.color_table_background().name());
-        style = style.arg(m_config.color_table_highlight().name());
-        style = style.arg(m_config.color_table_foreground().name());
-        ui->tableWidgetCalls->setStyleSheet(style);
-
-        // Set the table palette for inactive selected row
-        auto p = ui->tableWidgetCalls->palette();
-        p.setColor(QPalette::Highlight, m_config.color_table_highlight());
-        p.setColor(QPalette::HighlightedText,
-                   m_config.color_table_foreground());
-        p.setColor(QPalette::Inactive, QPalette::Highlight,
-                   p.color(QPalette::Active, QPalette::Highlight));
-        ui->tableWidgetCalls->setPalette(p);
-
-        // Set item fonts
-        for (int row = 0; row < ui->tableWidgetCalls->rowCount(); row++) {
-            auto bold = ui->tableWidgetCalls->item(row, 0)->text() == "\u2691";
-            for (int col = 0; col < ui->tableWidgetCalls->columnCount();
-                 col++) {
-                auto item = ui->tableWidgetCalls->item(row, col);
-                if (item) {
-                    auto f = m_config.table_font();
-                    if (bold) {
-                        f.setBold(true);
-                    }
-                    item->setFont(f);
-                }
-            }
-        }
-
-        // Column labels
-        ui->tableWidgetCalls->horizontalHeader()->setVisible(
-            showColumn("call", "labels"));
-
-        // Hide columns
-        ui->tableWidgetCalls->setColumnHidden(0, !showIconColumn);
-        ui->tableWidgetCalls->setColumnHidden(1,
-                                              !showColumn("call", "callsign"));
-        ui->tableWidgetCalls->setColumnHidden(2,
-                                              !showColumn("call", "timestamp"));
-        ui->tableWidgetCalls->setColumnHidden(3, !showColumn("call", "snr"));
-        ui->tableWidgetCalls->setColumnHidden(4, !showColumn("call", "offset"));
-        ui->tableWidgetCalls->setColumnHidden(
-            5, !showColumn("call", "tdrift", false));
-        ui->tableWidgetCalls->setColumnHidden(
-            6, !showColumn("call", "submode", false));
-        ui->tableWidgetCalls->setColumnHidden(
-            7, !showColumn("call", "grid", false));
-        ui->tableWidgetCalls->setColumnHidden(
-            8, !showColumn("call", "distance", false));
-        ui->tableWidgetCalls->setColumnHidden(
-            9, !showColumn("call", "azimuth", false));
-        ui->tableWidgetCalls->setColumnHidden(10, !showColumn("call", "log"));
-        ui->tableWidgetCalls->setColumnHidden(11,
-                                              !showColumn("call", "logName"));
-        ui->tableWidgetCalls->setColumnHidden(
-            12, !showColumn("call", "logComment"));
-
-        // Resize the table columns
-        ui->tableWidgetCalls->resizeColumnToContents(0);
-        ui->tableWidgetCalls->resizeColumnToContents(1);
-        ui->tableWidgetCalls->resizeColumnToContents(2);
-        ui->tableWidgetCalls->resizeColumnToContents(3);
-        ui->tableWidgetCalls->resizeColumnToContents(4);
-        ui->tableWidgetCalls->resizeColumnToContents(5);
-        ui->tableWidgetCalls->resizeColumnToContents(6);
-        ui->tableWidgetCalls->resizeColumnToContents(7);
-        ui->tableWidgetCalls->resizeColumnToContents(8);
-        ui->tableWidgetCalls->resizeColumnToContents(9);
-        ui->tableWidgetCalls->resizeColumnToContents(10);
-        ui->tableWidgetCalls->resizeColumnToContents(11);
-
-        // Reset the scroll position
-        ui->tableWidgetCalls->verticalScrollBar()->setValue(currentScrollPos);
-    }
-    ui->tableWidgetCalls->setUpdatesEnabled(true);
-}
+void displayCallActivity(); // JS8_Mainwindow/displayCallActivity.cpp
 
 void MainWindow::emitPTT(bool on) {
     qCDebug(mainwindow_js8) << "Setting PTT to" << (on ? "on" : "off");
@@ -11774,372 +7940,7 @@ void MainWindow::tcpNetworkMessage(Message const &message) {
     networkMessage(message);
 }
 
-void MainWindow::networkMessage(Message const &message) {
-    auto type = message.type();
-
-    if (type == "PING") {
-        return;
-    }
-
-    auto id = message.id();
-
-    qCDebug(mainwindow_js8) << "try processing network message" << type << id;
-
-    // Inspired by FLDigi
-    // TODO: MAIN.RX - Turn on RX
-    // TODO: MAIN.TX - Transmit
-    // TODO: MAIN.PTT - PTT
-    // TODO: MAIN.TUNE - Tune
-    // TODO: MAIN.HALT - Halt
-    // TODO: MAIN.AUTO - Auto
-    // TODO: MAIN.SPOT - Spot
-    // TODO: MAIN.HB - HB
-
-    // RIG.GET_FREQ - Get the current Frequency
-    // RIG.SET_FREQ - Set the current Frequency
-    if (type == "RIG.GET_FREQ") {
-        // Send WSJT-X Status message if protocol is enabled
-        if (m_wsjtxMessageMapper && m_config.wsjtx_protocol_enabled()) {
-            QString dx_call = callsignSelected();
-            QString dx_grid = "";
-            if (!dx_call.isEmpty() && m_callActivity.contains(dx_call)) {
-                dx_grid = m_callActivity[dx_call].grid;
-            }
-            QString tx_message = m_transmitting ? m_currentMessage : "";
-
-            m_wsjtxMessageMapper->sendStatusUpdate(
-                dialFrequency(), freq(),
-                "JS8", // mode
-                dx_call, m_config.my_callsign(), m_config.my_grid(), dx_grid,
-                true, // tx_enabled
-                m_transmitting,
-                m_decoderBusy || m_monitoring, // decoding
-                tx_message);
-        }
-
-        // Send native JSON message only if not conflicting with WSJT-X
-        bool skip_json = false;
-        if (m_config.wsjtx_protocol_enabled() &&
-            m_config.wsjtx_server_port() == m_config.udp_server_port() &&
-            m_config.wsjtx_server_name() == m_config.udp_server_name()) {
-            skip_json = true;
-        }
-
-        if (!skip_json) {
-            sendNetworkMessage(
-                "RIG.FREQ", "",
-                {{"_ID", id},
-                 {"FREQ", QVariant((quint64)dialFrequency() + freq())},
-                 {"DIAL", QVariant((quint64)dialFrequency())},
-                 {"OFFSET", QVariant((quint64)freq())}});
-        }
-        return;
-    }
-
-    if (type == "RIG.SET_FREQ") {
-        auto params = message.params();
-        if (params.contains("DIAL")) {
-            bool ok = false;
-            auto f = params["DIAL"].toInt(&ok);
-            if (ok) {
-                setRig(f);
-                displayDialFrequency();
-            }
-        }
-        if (params.contains("OFFSET")) {
-            bool ok = false;
-            auto f = params["OFFSET"].toInt(&ok);
-            if (ok) {
-                setFreqOffsetForRestore(f, false);
-            }
-        }
-    }
-
-    // STATION.GET_CALLSIGN - Get the current callsign
-    // STATION.GET_GRID - Get the current grid locator
-    // STATION.SET_GRID - Set the current grid locator
-    // STATION.GET_INFO - Get the current station qth
-    // STATION.SET_INFO - Set the current station qth
-    if (type == "STATION.GET_CALLSIGN") {
-        sendNetworkMessage("STATION.CALLSIGN", m_config.my_callsign(),
-                           {
-                               {"_ID", id},
-                           });
-        return;
-    }
-
-    if (type == "STATION.GET_GRID") {
-        sendNetworkMessage("STATION.GRID", m_config.my_grid(),
-                           {
-                               {"_ID", id},
-                           });
-        return;
-    }
-
-    if (type == "STATION.SET_GRID") {
-        m_config.set_dynamic_location(message.value());
-        sendNetworkMessage("STATION.GRID", m_config.my_grid(),
-                           {
-                               {"_ID", id},
-                           });
-        return;
-    }
-
-    if (type == "STATION.GET_INFO") {
-        sendNetworkMessage("STATION.INFO", m_config.my_info(),
-                           {
-                               {"_ID", id},
-                           });
-        return;
-    }
-
-    if (type == "STATION.SET_INFO") {
-        m_config.set_dynamic_station_info(message.value());
-        sendNetworkMessage("STATION.INFO", m_config.my_info(),
-                           {
-                               {"_ID", id},
-                           });
-        return;
-    }
-
-    if (type == "STATION.GET_STATUS") {
-        sendNetworkMessage("STATION.STATUS", m_config.my_status(),
-                           {
-                               {"_ID", id},
-                           });
-        return;
-    }
-
-    if (type == "STATION.SET_STATUS") {
-        m_config.set_dynamic_station_status(message.value());
-        sendNetworkMessage("STATION.STATUS", m_config.my_status(),
-                           {
-                               {"_ID", id},
-                           });
-        return;
-    }
-
-    // RX.GET_CALL_ACTIVITY
-    // RX.GET_CALL_SELECTED
-    // RX.GET_BAND_ACTIVITY
-    // RX.GET_TEXT
-
-    if (type == "RX.GET_CALL_ACTIVITY") {
-        auto now = DriftingDateTime::currentDateTimeUtc();
-        int callsignAging = m_config.callsign_aging();
-        QVariantMap calls = {
-            {"_ID", id},
-        };
-
-        foreach (auto cd, m_callActivity.values()) {
-            if (callsignAging &&
-                cd.utcTimestamp.secsTo(now) / 60 >= callsignAging) {
-                continue;
-            }
-            QVariantMap detail;
-            detail["SNR"] = QVariant(cd.snr);
-            detail["GRID"] = QVariant(cd.grid);
-            detail["UTC"] = QVariant(cd.utcTimestamp.toMSecsSinceEpoch());
-            calls[cd.call] = QVariant(detail);
-        }
-
-        sendNetworkMessage("RX.CALL_ACTIVITY", "", calls);
-        return;
-    }
-
-    if (type == "RX.GET_CALL_SELECTED") {
-        sendNetworkMessage("RX.CALL_SELECTED", callsignSelected(),
-                           {
-                               {"_ID", id},
-                           });
-        return;
-    }
-
-    if (type == "RX.GET_BAND_ACTIVITY") {
-        QVariantMap offsets = {
-            {"_ID", id},
-        };
-        for (auto const [offset, activity] : m_bandActivity.asKeyValueRange()) {
-            if (activity.isEmpty())
-                continue;
-
-            auto const d = activity.last();
-
-            offsets[QString("%1").arg(offset)] = QVariant(QVariantMap{
-                {"FREQ", QVariant(d.dial + d.offset)},
-                {"DIAL", QVariant(d.dial)},
-                {"OFFSET", QVariant(d.offset)},
-                {"TEXT", QVariant(d.text)},
-                {"SNR", QVariant(d.snr)},
-                {"UTC", QVariant(d.utcTimestamp.toMSecsSinceEpoch())}});
-        }
-
-        sendNetworkMessage("RX.BAND_ACTIVITY", "", offsets);
-        return;
-    }
-
-    if (type == "RX.GET_TEXT") {
-        sendNetworkMessage("RX.TEXT", ui->textEditRX->toPlainText().right(1024),
-                           {
-                               {"_ID", id},
-                           });
-        return;
-    }
-
-    // TX.GET_TEXT
-    // TX.SET_TEXT
-    // TX.SEND_MESSAGE
-
-    if (type == "TX.GET_TEXT") {
-        sendNetworkMessage("TX.TEXT",
-                           ui->extFreeTextMsgEdit->toPlainText().right(1024),
-                           {
-                               {"_ID", id},
-                           });
-        return;
-    }
-
-    if (type == "TX.SET_TEXT") {
-        addMessageText(message.value(), true);
-        sendNetworkMessage("TX.TEXT",
-                           ui->extFreeTextMsgEdit->toPlainText().right(1024),
-                           {
-                               {"_ID", id},
-                           });
-        return;
-    }
-
-    if (type == "TX.SEND_MESSAGE") {
-        auto text = message.value();
-        if (!text.isEmpty()) {
-            enqueueMessage(PriorityNormal, text, -1, nullptr);
-            processTxQueue();
-            return;
-        }
-    }
-
-    // MODE.GET_SPEED
-    // MODE.SET_SPEED
-    if (type == "MODE.GET_SPEED") {
-        sendNetworkMessage("MODE.SPEED", "",
-                           {
-                               {"_ID", id},
-                               {"SPEED", m_nSubMode},
-                           });
-        return;
-    }
-
-    if (type == "MODE.SET_SPEED") {
-        auto ok = false;
-        auto const speed =
-            message.params().value("SPEED", QVariant(m_nSubMode)).toInt(&ok);
-        if (ok) {
-            if (speed == Varicode::JS8CallNormal)
-                ui->actionModeJS8Normal->setChecked(true);
-            else if (speed == Varicode::JS8CallFast)
-                ui->actionModeJS8Fast->setChecked(true);
-            else if (speed == Varicode::JS8CallTurbo)
-                ui->actionModeJS8Turbo->setChecked(true);
-            else if (speed == Varicode::JS8CallSlow)
-                ui->actionModeJS8Slow->setChecked(true);
-            else if (speed == Varicode::JS8CallUltra)
-                ui->actionModeJS8Ultra->setChecked(true);
-            setupJS8();
-        }
-        sendNetworkMessage("MODE.SPEED", "",
-                           {
-                               {"_ID", id},
-                               {"SPEED", m_nSubMode},
-                           });
-        return;
-    }
-
-    // INBOX.GET_MESSAGES
-    // INBOX.STORE_MESSAGE
-    if (type == "INBOX.GET_MESSAGES") {
-        QString selectedCall =
-            message.params().value("CALLSIGN", "").toString();
-        if (selectedCall.isEmpty()) {
-            selectedCall = "%";
-        }
-
-        Inbox inbox(inboxPath());
-        if (!inbox.open()) {
-            return;
-        }
-
-        QList<QPair<int, Message>> msgs;
-        msgs.append(
-            inbox.values("STORE", "$.params.TO", selectedCall, 0, 1000));
-        msgs.append(
-            inbox.values("READ", "$.params.FROM", selectedCall, 0, 1000));
-        foreach (auto pair, inbox.values("UNREAD", "$.params.FROM",
-                                         selectedCall, 0, 1000)) {
-            msgs.append(pair);
-        }
-        std::stable_sort(
-            msgs.begin(), msgs.end(),
-            [](QPair<int, Message> const &a, QPair<int, Message> const &b) {
-                return QVariant::compare(a.second.params().value("UTC"),
-                                         b.second.params().value("UTC")) ==
-                       QPartialOrdering::Greater;
-            });
-
-        QVariantList l;
-        foreach (auto pair, msgs) {
-            l << pair.second.toVariantMap();
-        }
-
-        sendNetworkMessage("INBOX.MESSAGES", "",
-                           {
-                               {"_ID", id},
-                               {"MESSAGES", l},
-                           });
-        return;
-    }
-
-    if (type == "INBOX.STORE_MESSAGE") {
-        QString selectedCall =
-            message.params().value("CALLSIGN", "").toString();
-        if (selectedCall.isEmpty()) {
-            return;
-        }
-
-        QString text = message.params().value("TEXT", "").toString();
-        if (text.isEmpty()) {
-            return;
-        }
-
-        CommandDetail d = {};
-        d.cmd = " MSG ";
-        d.to = selectedCall;
-        d.from = m_config.my_callsign();
-        d.relayPath = d.from;
-        d.text = text;
-        d.utcTimestamp = DriftingDateTime::currentDateTimeUtc();
-        d.submode = m_nSubMode;
-
-        auto mid = addCommandToStorage("STORE", d);
-
-        sendNetworkMessage("INBOX.MESSAGE", "",
-                           {
-                               {"_ID", id},
-                               {"ID", mid},
-                           });
-        return;
-    }
-
-    // WINDOW.RAISE
-
-    if (type == "WINDOW.RAISE") {
-        setWindowState(Qt::WindowActive);
-        activateWindow();
-        raise();
-        return;
-    }
-
-    qCDebug(mainwindow_js8) << "Unable to process networkMessage:" << type;
-}
+void networkMessage(); // JS8_Mainwindow/networkMessage.cpp
 
 bool MainWindow::canSendNetworkMessage() {
     return m_config.udpEnabled() || m_config.tcpEnabled();
@@ -12226,8 +8027,8 @@ void MainWindow::statusUpdate() {
             dialFrequency(), freq(),
             "JS8", // mode
             dx_call, m_config.my_callsign(), m_config.my_grid(), dx_grid,
-            true, // tx_enabled - JS8Call always allows TX when not in special
-                  // modes
+            true, // tx_enabled - JS8Call always allows TX when not in
+                  // special modes
             m_transmitting,
             m_decoderBusy || m_monitoring, // decoding
             tx_message);
@@ -12273,7 +8074,8 @@ void MainWindow::childEvent(QChildEvent *e) {
 }
 
 // add widget and any child widgets to our event filter so that we can
-// take action on key press ad mouse press events anywhere in the main window
+// take action on key press ad mouse press events anywhere in the main
+// window
 void MainWindow::add_child_to_event_filter(QObject *target) {
     if (target && target->isWidgetType()) {
         target->installEventFilter(this);
